refactor(torghut): split autonomy lane ownership
