+++ services/torghut/app/trading/autonomy/lane.py
@@ -10,7 +10,7 @@ from dataclasses import dataclass, field
 from datetime import datetime, timezone
 from decimal import Decimal
 from pathlib import Path
-from typing import Any, Callable, Iterable, Mapping, Sequence, cast
+from typing import Any, Callable, Mapping, Sequence, cast
 
 from sqlalchemy import delete, select
 from sqlalchemy.orm import Session
@@ -136,405 +136,97 @@ from .phase_manifest_contract import (
     coerce_phase_status,
     normalize_phase_transitions,
 )
-
-
-_AUTONOMY_PHASE_ORDER: tuple[str, ...] = AUTONOMY_PHASE_ORDER
-
-
-_ACTUATION_INTENT_SCHEMA_VERSION = "torghut.autonomy.actuation-intent.v1"
-_ACTUATION_CONFIRMATION_PHRASE = "ACTUATE_TORGHUT"
-_ACTUATION_INTENT_PATH = "gates/actuation-intent.json"
-_AUTONOMY_LANE_SCHEMA_VERSION = "torghut-autonomy-stage-manifest-v1"
-_PROFITABILITY_STAGE_MANIFEST_SCHEMA_VERSION = "profitability-stage-manifest-v1"
-_PROFITABILITY_STAGE_MANIFEST_PATH = (
-    "profitability/profitability-stage-manifest-v1.json"
+from .lane_candidate_payloads import (
+    build_candidate_alpha_readiness_payload as _build_candidate_alpha_readiness_payload,
+    build_candidate_state_payload as _build_candidate_state_payload,
+    is_runbook_valid as _is_runbook_valid,
+    normalize_strategy_artifacts as _normalize_strategy_artifacts,
 )
-_STAGE_CANDIDATE_GENERATION = "candidate-generation"
-_STAGE_EVALUATION = "evaluation"
-_STAGE_RECOMMENDATION = "promotion-recommendation"
-_STRESS_METRICS_ARTIFACT_PATH = "stress-metrics-v1.json"
-_CONTAMINATION_REGISTRY_ARTIFACT_PATH = "contamination-leakage-report-v1.json"
-_HMM_STATE_POSTERIOR_ARTIFACT_PATH = "gates/hmm-state-posterior-v1.json"
-_EXPERT_ROUTER_REGISTRY_ARTIFACT_PATH = "gates/expert-router-registry-v1.json"
-_FOLD_METRICS_ARTIFACT_PATH = "fold-metrics-v1.json"
-_STRESS_METRICS_CASES = ("spread", "volatility", "liquidity", "halt")
-_BENCHMARK_PARITY_REPORT_PATH = "gates/benchmark-parity-report-v1.json"
-_FOUNDATION_ROUTER_PARITY_REPORT_PATH = "router/foundation-router-parity-report-v1.json"
-_DEEPLOB_BDLOB_REPORT_PATH = "microstructure/deeplob-bdlob-report-v1.json"
-_ADVISOR_FALLBACK_SLO_REPORT_PATH = "execution/advisor-fallback-slo-report-v1.json"
-_STAGE_PROFITABILITY = "profitability_stage_manifest"
-_V6_08_GOVERNING_DESIGN_DOC = "docs/torghut/design-system/v6/08-profitability-research-validation-execution-governance-system.md"
-
-
-@dataclass(frozen=True)
-class _StageManifestRecord:
-    stage: str
-    stage_index: int
-    stage_trace_id: str
-    lineage_hash: str
-    artifact_hashes: dict[str, str]
-    stage_payload_hash: str
-    created_at: str
-    created_by: str = "run_autonomous_lane"
-    parent_lineage_hash: str | None = None
-    parent_stage: str | None = None
-    inputs: dict[str, str] = field(default_factory=lambda: cast(dict[str, str], {}))
-
-
-def _coerce_evidence_bool(value: object) -> bool | None:
-    if isinstance(value, bool):
-        return value
-    if isinstance(value, int | float):
-        if not math.isfinite(value):
-            return None
-        return value != 0
-    if value is None:
-        return None
-    normalized = str(value).strip().lower()
-    if normalized in {"true", "1", "yes", "on"}:
-        return True
-    if normalized in {"false", "0", "no", "off", ""}:
-        return False
-    return None
-
-
-def _normalize_strategy_artifacts(value: object) -> object | None:
-    if not isinstance(value, dict):
-        return value
-
-    payload = cast(dict[str, object], value)
-    kind = str(payload.get("kind", "")).strip().lower()
-    if kind == "configmap":
-        data = payload.get("data")
-        if not isinstance(data, dict):
-            return None
-        data_payload = cast(dict[str, object], data)
-        strategies_yaml = data_payload.get("strategies.yaml")
-        if not isinstance(strategies_yaml, str):
-            return None
-        try:
-            nested_payload = cast(object, yaml.safe_load(strategies_yaml))
-        except Exception:
-            return None
-        return _normalize_strategy_artifacts(nested_payload)
-
-    return payload
-
-
-def _is_runbook_valid(strategy_configmap_path: Path | None) -> bool:
-    resolved_path = strategy_configmap_path or _default_strategy_configmap_path()
-    if not resolved_path.exists() or not resolved_path.is_file():
-        return False
-
-    try:
-        raw_payload = cast(
-            object, yaml.safe_load(resolved_path.read_text(encoding="utf-8"))
-        )
-    except Exception:
-        return False
-
-    normalized_payload = _normalize_strategy_artifacts(raw_payload)
-
-    if normalized_payload is None:
-        return False
-
-    if isinstance(normalized_payload, dict):
-        payload = cast(dict[str, object], normalized_payload)
-        strategies = payload.get("strategies")
-    elif isinstance(normalized_payload, list):
-        strategies = cast(list[object], normalized_payload)
-    else:
-        return False
-
-    if not isinstance(strategies, list):
-        return False
-
-    strategy_items = cast(list[object], strategies)
-    return bool(strategy_items) and all(
-        isinstance(item, dict) for item in strategy_items
-    )
-
-
-def _build_candidate_state_payload(
-    *,
-    candidate_id: str,
-    run_id: str,
-    promotion_target: PromotionTarget,
-    approval_token: str | None,
-    runtime_strategies: list[StrategyRuntimeConfig],
-    now: datetime,
-    code_version: str,
-    runbook_validated: bool,
-    dependency_quorum_payload: dict[str, Any],
-    alpha_readiness_payload: dict[str, Any],
-) -> dict[str, Any]:
-    return {
-        "candidateId": candidate_id,
-        "runId": run_id,
-        "activeStage": "gate-evaluation",
-        "paused": False,
-        "datasetSnapshotRef": "signals_window",
-        "noSignalReason": None,
-        "runbookValidated": runbook_validated,
-        "dependencyQuorum": dependency_quorum_payload,
-        "alphaReadiness": alpha_readiness_payload,
-        "rollbackReadiness": {
-            "killSwitchDryRunPassed": bool(runtime_strategies),
-            "gitopsRevertDryRunPassed": (
-                promotion_target != "live" or bool(approval_token)
-            ),
-            "strategyDisableDryRunPassed": bool(runtime_strategies),
-            "dryRunCompletedAt": now.isoformat(),
-            "humanApproved": promotion_target != "live" or bool(approval_token),
-            "rollbackTarget": f"{code_version or 'unknown'}",
-        },
-    }
-
-
-def _build_candidate_alpha_readiness_payload(
-    *,
-    runtime_strategies: Sequence[StrategyRuntimeConfig],
-) -> tuple[dict[str, Any], dict[str, Any]]:
-    registry = load_hypothesis_registry()
-    dependency_quorum = resolve_hypothesis_dependency_quorum(registry)
-    strategy_families = sorted(
-        {
-            str(strategy.strategy_type).strip()
-            for strategy in runtime_strategies
-            if str(strategy.strategy_type).strip()
-        }
-    )
-    matched_hypotheses = [
-        manifest
-        for manifest in registry.items
-        if manifest.strategy_family in strategy_families
-    ]
-    matched_strategy_families = {
-        manifest.strategy_family for manifest in matched_hypotheses
-    }
-    missing_strategy_families = sorted(
-        family
-        for family in strategy_families
-        if family not in matched_strategy_families
-    )
-    reasons: list[str] = []
-    if not registry.loaded:
-        reasons.append("hypothesis_registry_unavailable")
-    if registry.errors:
-        reasons.append("hypothesis_registry_errors_present")
-    if missing_strategy_families:
-        reasons.append("strategy_family_hypothesis_unmapped")
-    if dependency_quorum.decision != "allow":
-        reasons.append(f"jangar_dependency_quorum_{dependency_quorum.decision}")
-    return (
-        {
-            "mode": "candidate_alignment_v1",
-            "registry_loaded": registry.loaded,
-            "registry_path": registry.path,
-            "registry_errors": list(registry.errors),
-            "strategy_families": strategy_families,
-            "matched_hypothesis_ids": sorted(
-                {manifest.hypothesis_id for manifest in matched_hypotheses}
-            ),
-            "missing_strategy_families": missing_strategy_families,
-            "promotion_eligible": (
-                registry.loaded
-                and not registry.errors
-                and not missing_strategy_families
-                and dependency_quorum.decision == "allow"
-            ),
-            "reasons": reasons,
-            "dependency_quorum": dependency_quorum.as_payload(),
-        },
-        dependency_quorum.as_payload(),
-    )
-
-
-@dataclass(frozen=True)
-class AutonomousLaneResult:
-    run_id: str
-    candidate_id: str
-    output_dir: Path
-    gate_report_path: Path
-    actuation_intent_path: Path | None
-    paper_patch_path: Path | None
-    phase_manifest_path: Path
-    benchmark_parity_path: Path
-    foundation_router_parity_path: Path
-    gate_report_trace_id: str
-    recommendation_trace_id: str
-    recommendation_artifact_path: Path
-    candidate_spec_path: Path
-    candidate_generation_manifest_path: Path
-    evaluation_manifest_path: Path
-    recommendation_manifest_path: Path
-    profitability_manifest_path: Path
-    stage_trace_ids: dict[str, str] = field(
-        default_factory=lambda: cast(dict[str, str], {})
-    )
-    stage_lineage_root: str | None = None
-
-
-@dataclass(frozen=True)
-class _StrategyFactoryBridge:
-    result: AlphaLaneResult
-    candidate_spec_payload: dict[str, Any]
-    evaluation_payload: dict[str, Any]
-    recommendation_payload: dict[str, Any]
-    attempt_payload: dict[str, Any]
-    validation_payloads: dict[str, dict[str, Any]]
-    sequential_trial_payload: dict[str, Any]
-    cost_calibration_payload: dict[str, Any]
-
-
-def _load_price_frame(path: Path) -> pd.DataFrame:
-    return pd.read_csv(path, index_col=0, parse_dates=True)
-
-
-def _build_strategy_factory_bridge(
-    *,
-    output_dir: Path,
-    notes_artifact_root: str | None,
-    train_prices_path: Path | None,
-    test_prices_path: Path | None,
-    alpha_gate_policy_path: Path | None,
-    repository: str | None,
-    base: str | None,
-    head: str | None,
-    priority_id: str | None,
-    promotion_target: str,
-    now: datetime,
-) -> _StrategyFactoryBridge | None:
-    if train_prices_path is None or test_prices_path is None:
-        return None
-
-    bridge_result = run_alpha_discovery_lane(
-        artifact_path=output_dir / "strategy-factory",
-        train_prices=_load_price_frame(train_prices_path),
-        test_prices=_load_price_frame(test_prices_path),
-        repository=repository,
-        base=base,
-        head=head,
-        priority_id=priority_id,
-        notes_artifact_path=notes_artifact_root,
-        gate_policy_path=alpha_gate_policy_path,
-        promotion_target=promotion_target,
-        evaluated_at=now,
-        persist_results=False,
-    )
-    validation_payloads: dict[str, dict[str, Any]] = {}
-    for test_name, artifact_path in bridge_result.validation_artifact_paths.items():
-        validation_payload = _load_json_if_exists(artifact_path)
-        if validation_payload is None:
-            continue
-        validation_payloads[test_name] = validation_payload
-    return _StrategyFactoryBridge(
-        result=bridge_result,
-        candidate_spec_payload=_load_json_if_exists(bridge_result.candidate_spec_path)
-        or {},
-        evaluation_payload=_load_json_if_exists(bridge_result.evaluation_report_path)
-        or {},
-        recommendation_payload=_load_json_if_exists(
-            bridge_result.recommendation_artifact_path
-        )
-        or {},
-        attempt_payload=_load_json_if_exists(bridge_result.attempt_ledger_path) or {},
-        validation_payloads=validation_payloads,
-        sequential_trial_payload=_load_json_if_exists(
-            bridge_result.sequential_trial_path
-        )
-        or {},
-        cost_calibration_payload=_load_json_if_exists(
-            bridge_result.cost_calibration_path
-        )
-        or {},
-    )
-
-
-def _strategy_factory_gate_summary(
-    bridge: _StrategyFactoryBridge,
-    *,
-    promotion_target: str,
-) -> tuple[bool, list[str], dict[str, Any]]:
-    reasons: list[str] = []
-    recommendation = (
-        cast(dict[str, Any], bridge.recommendation_payload.get("recommendation"))
-        if isinstance(bridge.recommendation_payload.get("recommendation"), dict)
-        else {}
-    )
-    strategy_payload = (
-        cast(dict[str, Any], bridge.candidate_spec_payload.get("strategy_factory"))
-        if isinstance(bridge.candidate_spec_payload.get("strategy_factory"), dict)
-        else {}
-    )
-    economic_card = (
-        cast(dict[str, Any], strategy_payload.get("economic_validity_card"))
-        if isinstance(strategy_payload.get("economic_validity_card"), dict)
-        else {}
-    )
-    comparator = (
-        cast(dict[str, Any], strategy_payload.get("null_comparator_summary"))
-        if isinstance(strategy_payload.get("null_comparator_summary"), dict)
-        else {}
-    )
-    sequential = bridge.sequential_trial_payload
-    calibration = bridge.cost_calibration_payload
-
-    if not bool(recommendation.get("eligible")):
-        reasons.append("strategy_factory_recommendation_not_eligible")
-    if str(economic_card.get("status") or "").strip().lower() != "pass":
-        reasons.append("strategy_factory_economic_validity_failed")
-    if not bool(comparator.get("baseline_outperformed")):
-        reasons.append("strategy_factory_baseline_not_outperformed")
-
-    sequential_status = str(sequential.get("status") or "").strip().lower()
-    calibration_status = str(calibration.get("status") or "").strip().lower()
-    if promotion_target in {"paper", "live"} and sequential_status not in {
-        "paper_ready",
-        "paper_only",
-    }:
-        reasons.append("strategy_factory_sequential_not_ready")
-    if promotion_target == "live":
-        if sequential_status != "paper_ready":
-            reasons.append("strategy_factory_live_requires_paper_ready")
-        if calibration_status != "calibrated":
-            reasons.append("strategy_factory_live_requires_calibrated_costs")
-
-    summary = {
-        "allowed": len(reasons) == 0,
-        "reasons": reasons,
-        "candidate_spec_artifact": str(bridge.result.candidate_spec_path),
-        "evaluation_artifact": str(bridge.result.evaluation_report_path),
-        "recommendation_artifact": str(bridge.result.recommendation_artifact_path),
-        "attempt_ledger_artifact": str(bridge.result.attempt_ledger_path),
-        "sequential_trial_artifact": str(bridge.result.sequential_trial_path),
-        "cost_calibration_artifact": str(bridge.result.cost_calibration_path),
-        "sequential_status": sequential_status or None,
-        "cost_calibration_status": calibration_status or None,
-        "baseline_outperformed": comparator.get("baseline_outperformed"),
-        "posterior_edge_summary": strategy_payload.get("posterior_edge_summary"),
-    }
-    return len(reasons) == 0, reasons, summary
-
-
-def _strategy_factory_artifact_refs(
-    bridge: _StrategyFactoryBridge | None,
-) -> list[str]:
-    if bridge is None:
-        return []
-    refs = [
-        bridge.result.candidate_spec_path,
-        bridge.result.evaluation_report_path,
-        bridge.result.recommendation_artifact_path,
-        bridge.result.attempt_ledger_path,
-        bridge.result.sequential_trial_path,
-        bridge.result.cost_calibration_path,
-        bridge.result.candidate_generation_manifest_path,
-        bridge.result.evaluation_manifest_path,
-        bridge.result.recommendation_manifest_path,
-        *bridge.result.validation_artifact_paths.values(),
-    ]
-    return [str(path) for path in refs]
+from .lane_common import (
+    ACTUATION_CONFIRMATION_PHRASE as _ACTUATION_CONFIRMATION_PHRASE,
+    ACTUATION_INTENT_PATH as _ACTUATION_INTENT_PATH,
+    ACTUATION_INTENT_SCHEMA_VERSION as _ACTUATION_INTENT_SCHEMA_VERSION,
+    ADVISOR_FALLBACK_SLO_REPORT_PATH as _ADVISOR_FALLBACK_SLO_REPORT_PATH,
+    AUTONOMY_LANE_SCHEMA_VERSION as _AUTONOMY_LANE_SCHEMA_VERSION,
+    AutonomousLaneResult,
+    BENCHMARK_PARITY_REPORT_PATH as _BENCHMARK_PARITY_REPORT_PATH,
+    CONTAMINATION_REGISTRY_ARTIFACT_PATH as _CONTAMINATION_REGISTRY_ARTIFACT_PATH,
+    DEEPLOB_BDLOB_REPORT_PATH as _DEEPLOB_BDLOB_REPORT_PATH,
+    EXPERT_ROUTER_REGISTRY_ARTIFACT_PATH as _EXPERT_ROUTER_REGISTRY_ARTIFACT_PATH,
+    FOLD_METRICS_ARTIFACT_PATH as _FOLD_METRICS_ARTIFACT_PATH,
+    FOUNDATION_ROUTER_PARITY_REPORT_PATH as _FOUNDATION_ROUTER_PARITY_REPORT_PATH,
+    HMM_STATE_POSTERIOR_ARTIFACT_PATH as _HMM_STATE_POSTERIOR_ARTIFACT_PATH,
+    LANE_AUTONOMY_PHASE_ORDER as _AUTONOMY_PHASE_ORDER,
+    PROFITABILITY_STAGE_MANIFEST_PATH as _PROFITABILITY_STAGE_MANIFEST_PATH,
+    PROFITABILITY_STAGE_MANIFEST_SCHEMA_VERSION as _PROFITABILITY_STAGE_MANIFEST_SCHEMA_VERSION,
+    STAGE_CANDIDATE_GENERATION as _STAGE_CANDIDATE_GENERATION,
+    STAGE_EVALUATION as _STAGE_EVALUATION,
+    STAGE_PROFITABILITY as _STAGE_PROFITABILITY,
+    STAGE_RECOMMENDATION as _STAGE_RECOMMENDATION,
+    STRESS_METRICS_ARTIFACT_PATH as _STRESS_METRICS_ARTIFACT_PATH,
+    STRESS_METRICS_CASES as _STRESS_METRICS_CASES,
+    StageManifestRecord as _StageManifestRecord,
+    V6_08_GOVERNING_DESIGN_DOC as _V6_08_GOVERNING_DESIGN_DOC,
+    as_object_dict as _as_object_dict,
+    coerce_evidence_bool as _coerce_evidence_bool,
+    default_strategy_configmap_path as _default_strategy_configmap_path,
+    ensure_utc as _ensure_utc,
+    safe_int as _safe_int,
+    sha256_path as _sha256_path,
+    stable_hash as _stable_hash,
+)
+from .lane_profitability_manifest import (
+    build_profitability_stage_manifest as _build_profitability_stage_manifest,
+)
+from .lane_phase_payloads import (
+    build_actuation_intent_payload as _build_actuation_intent_payload,
+    build_phase_manifest as _build_phase_manifest,
+    candidate_state_readiness_payload as _candidate_state_readiness_payload,
+    coalesce_governance_context as _coalesce_governance_context,
+    coerce_gate_phase_gates as _coerce_gate_phase_gates,
+    coerce_int as _coerce_int,
+    coerce_path_strings as _coerce_path_strings,
+    coerce_str as _coerce_str,
+    normalize_governance_inputs as _normalize_governance_inputs,
+    prepare_lane_output_dirs as _prepare_lane_output_dirs,
+)
+from .lane_regime_artifacts import (
+    build_contamination_registry_payload as _build_contamination_registry_payload,
+    build_expert_router_registry_payload as _build_expert_router_registry_payload,
+    build_expert_router_weights as _build_expert_router_weights,
+    build_hmm_state_posterior_payload as _build_hmm_state_posterior_payload,
+    decimal_or_none as _decimal_or_none,
+    decimal_or_zero as _decimal_or_zero,
+    normalize_expert_weights as _normalize_expert_weights,
+    normalize_hmm_regime_id as _normalize_hmm_regime_id,
+)
+from .lane_stage_artifacts import (
+    artifact_authority_for_check as _artifact_authority_for_check,
+    artifact_authority_for_evidence as _artifact_authority_for_evidence,
+    artifact_hashes as _artifact_hashes,
+    build_bridge_evidence_payload as _build_bridge_evidence_payload,
+    build_janus_q_summary_from_payloads as _build_janus_q_summary_from_payloads,
+    build_portfolio_promotion_summary as _build_portfolio_promotion_summary,
+    build_stage_lineage_payload as _build_stage_lineage_payload,
+    build_vnext_gate_summary as _build_vnext_gate_summary,
+    empirical_artifact_authority as _empirical_artifact_authority,
+    extract_janus_q_metrics as _extract_janus_q_metrics,
+    load_configured_empirical_payload as _load_configured_empirical_payload,
+    load_json_if_exists as _load_json_if_exists,
+    manifest_artifact_payload as _manifest_artifact_payload,
+    manifest_relative_path as _manifest_relative_path,
+    readable_notes_iteration_number as _readable_notes_iteration_number,
+    resolve_optional_service_path as _resolve_optional_service_path,
+    write_iteration_notes as _write_iteration_notes,
+    write_stage_manifest as _write_stage_manifest,
+)
+from .lane_strategy_factory import (
+    StrategyFactoryBridge as _StrategyFactoryBridge,
+    build_strategy_factory_bridge as _build_strategy_factory_bridge,
+    load_price_frame as _load_price_frame,
+    strategy_factory_artifact_refs as _strategy_factory_artifact_refs,
+    strategy_factory_gate_summary as _strategy_factory_gate_summary,
+)
 
 
 def upsert_autonomy_no_signal_run(
@@ -641,1452 +333,6 @@ def upsert_autonomy_no_signal_run(
             session.add(existing_run)
         session.commit()
         return run_id
-
-
-def _ensure_utc(value: datetime) -> datetime:
-    return value if value.tzinfo is not None else value.replace(tzinfo=timezone.utc)
-
-
-def _safe_int(value: Any) -> int:
-    try:
-        return int(value)
-    except (TypeError, ValueError):
-        return 0
-
-
-def _as_object_dict(value: object) -> dict[str, object]:
-    if not isinstance(value, Mapping):
-        return {}
-    mapping = cast(Mapping[object, object], value)
-    return {str(key): item for key, item in mapping.items()}
-
-
-def _stable_hash(payload: object) -> str:
-    payload_json = json.dumps(payload, sort_keys=True, separators=(",", ":"))
-    return hashlib.sha256(payload_json.encode("utf-8")).hexdigest()
-
-
-def _artifact_hashes(artifacts: Mapping[str, Path | None]) -> dict[str, str]:
-    digests: dict[str, str] = {}
-    for key, artifact_path in artifacts.items():
-        if artifact_path is not None and artifact_path.exists():
-            digests[key] = _sha256_path(artifact_path)
-    return digests
-
-
-def _readable_notes_iteration_number(notes_dir: Path) -> int:
-    pattern = re.compile(r"^iteration-(\d+)\.md$")
-    highest = 0
-    for item in notes_dir.glob("iteration-*.md"):
-        match = pattern.match(item.name)
-        if not match:
-            continue
-        try:
-            candidate = int(match.group(1))
-        except (TypeError, ValueError):
-            continue
-        if candidate > highest:
-            highest = candidate
-    return highest + 1
-
-
-def _write_stage_manifest(
-    *,
-    stage: str,
-    stage_index: int,
-    stage_output_dir: Path,
-    run_id: str,
-    candidate_id: str,
-    lineage_parent_hash: str | None,
-    lineage_parent_stage: str | None,
-    inputs: dict[str, str],
-    input_artifacts: Mapping[str, Path | None],
-    output_artifacts: Mapping[str, Path | None],
-    created_at: datetime,
-) -> _StageManifestRecord:
-    stage_manifest = {
-        "schema_version": _AUTONOMY_LANE_SCHEMA_VERSION,
-        "stage": stage,
-        "stage_index": stage_index,
-        "run_id": run_id,
-        "candidate_id": candidate_id,
-        "created_at": created_at.isoformat(),
-        "inputs": dict(inputs),
-        "input_artifacts": {
-            key: {
-                "path": str(path),
-                "sha256": _sha256_path(path),
-            }
-            for key, path in input_artifacts.items()
-            if path is not None
-        },
-        "output_artifacts": {
-            key: {
-                "path": str(path),
-                "sha256": _sha256_path(path),
-            }
-            for key, path in output_artifacts.items()
-            if path is not None
-        },
-        "parent_lineage_hash": lineage_parent_hash,
-        "parent_stage": lineage_parent_stage,
-    }
-    stage_payload_hash = _stable_hash(stage_manifest)
-    stage_trace_id = stage_payload_hash[:24]
-    artifact_hashes = _artifact_hashes(output_artifacts)
-    record = _StageManifestRecord(
-        stage=stage,
-        stage_index=stage_index,
-        stage_trace_id=stage_trace_id,
-        lineage_hash=stage_payload_hash,
-        artifact_hashes=artifact_hashes,
-        stage_payload_hash=stage_payload_hash,
-        created_at=created_at.isoformat(),
-        parent_lineage_hash=lineage_parent_hash,
-        parent_stage=lineage_parent_stage,
-        inputs=dict(inputs),
-    )
-    stage_manifest["stage_trace_id"] = stage_trace_id
-    stage_manifest["lineage_hash"] = stage_payload_hash
-    stage_manifest["artifact_hashes"] = artifact_hashes
-    manifest_path = stage_output_dir / f"{stage}-manifest.json"
-    stage_manifest["artifact_count"] = len(artifact_hashes)
-    manifest_path.write_text(json.dumps(stage_manifest, indent=2), encoding="utf-8")
-    return record
-
-
-def _build_stage_lineage_payload(
-    stage_records: Sequence[_StageManifestRecord],
-    manifest_paths: Mapping[str, Path],
-) -> dict[str, Any]:
-    stages_payload: dict[str, Any] = {}
-    for record in stage_records:
-        manifest_path = manifest_paths.get(record.stage)
-        stages_payload[record.stage] = {
-            "index": record.stage_index,
-            "stage_trace_id": record.stage_trace_id,
-            "lineage_hash": record.lineage_hash,
-            "parent_stage": record.parent_stage,
-            "parent_lineage_hash": record.parent_lineage_hash,
-            "stage_payload_hash": record.stage_payload_hash,
-            "created_at": record.created_at,
-            "created_by": record.created_by,
-            "inputs": record.inputs,
-            "artifact_hashes": record.artifact_hashes,
-            "manifest_path": str(manifest_path) if manifest_path else None,
-        }
-    return {
-        "schema_version": _AUTONOMY_LANE_SCHEMA_VERSION,
-        "root_lineage_hash": stage_records[0].lineage_hash if stage_records else None,
-        "stages": stages_payload,
-    }
-
-
-def _manifest_relative_path(artifact_root: Path, artifact_path: Path) -> str:
-    try:
-        return str(artifact_path.relative_to(artifact_root))
-    except ValueError:
-        return str(artifact_path)
-
-
-def _artifact_authority_for_check(
-    *,
-    stage_name: str,
-    check_name: str,
-) -> dict[str, Any] | None:
-    if check_name in {
-        "benchmark_parity_present",
-        "foundation_router_parity_present",
-        "janus_event_car_present",
-        "janus_hgrm_reward_present",
-        "expert_router_registry_present",
-        "hmm_state_posterior_present",
-    }:
-        return evidence_contract_payload(
-            provenance=ArtifactProvenance.SYNTHETIC_GENERATED,
-            maturity=EvidenceMaturity.STUB,
-            authoritative=False,
-            placeholder=True,
-            notes="Deterministic scaffold artifact; never promotion-authoritative.",
-        )
-    if check_name in {
-        "evaluation_report_present",
-        "walkforward_results_present",
-        "baseline_evaluation_report_present",
-        "profitability_benchmark_present",
-        "profitability_evidence_present",
-        "profitability_validation_present",
-        "contamination_registry_present",
-        "advisor_fallback_slo_present",
-        "recalibration_report_present",
-        "gate_report_present",
-    }:
-        return evidence_contract_payload(
-            provenance=ArtifactProvenance.HISTORICAL_MARKET_REPLAY,
-            maturity=EvidenceMaturity.UNCALIBRATED,
-            calibration_summary={
-                "status": "pending_calibration",
-                "stage": stage_name,
-                "check": check_name,
-            },
-        )
-    return None
-
-
-def _artifact_authority_for_evidence(name: str) -> dict[str, Any]:
-    evidence_name = name.strip().lower()
-    if evidence_name in {
-        "benchmark_parity",
-        "foundation_router_parity",
-        "janus_event_car",
-        "janus_hgrm_reward",
-        "janus_q",
-        "hmm_state_posterior",
-        "expert_router_registry",
-        "deeplob_bdlob_contract",
-    }:
-        return evidence_contract_payload(
-            provenance=ArtifactProvenance.SYNTHETIC_GENERATED,
-            maturity=EvidenceMaturity.STUB,
-            authoritative=False,
-            placeholder=True,
-            notes="Synthetic or deterministic scaffold evidence; blocked for promotion.",
-        )
-    return evidence_contract_payload(
-        provenance=ArtifactProvenance.HISTORICAL_MARKET_REPLAY,
-        maturity=EvidenceMaturity.UNCALIBRATED,
-        calibration_summary={
-            "status": "pending_calibration",
-            "evidence_name": evidence_name,
-        },
-    )
-
-
-def _empirical_artifact_authority(
-    *,
-    evidence_name: str,
-    provenance: ArtifactProvenance = ArtifactProvenance.HISTORICAL_MARKET_REPLAY,
-) -> dict[str, Any]:
-    return evidence_contract_payload(
-        provenance=provenance,
-        maturity=EvidenceMaturity.EMPIRICALLY_VALIDATED,
-        authoritative=True,
-        placeholder=False,
-        calibration_summary={
-            "status": "external_empirical_source",
-            "evidence_name": evidence_name,
-        },
-        notes="Artifact loaded from configured empirical source.",
-    )
-
-
-def _resolve_optional_service_path(raw_value: str | None) -> Path | None:
-    normalized = (raw_value or "").strip()
-    if not normalized:
-        return None
-    path = Path(normalized).expanduser()
-    if path.is_absolute():
-        return path
-    service_root = Path(__file__).resolve().parents[3]
-    return service_root / path
-
-
-def _load_configured_empirical_payload(
-    *,
-    path_value: str | None,
-    expected_schema_version: str | None,
-    evidence_name: str,
-    provenance: ArtifactProvenance = ArtifactProvenance.HISTORICAL_MARKET_REPLAY,
-) -> dict[str, Any] | None:
-    resolved = _resolve_optional_service_path(path_value)
-    if resolved is None or not resolved.exists() or not resolved.is_file():
-        return None
-    raw_payload = json.loads(resolved.read_text(encoding="utf-8"))
-    if not isinstance(raw_payload, dict):
-        raise RuntimeError(f"configured_empirical_payload_invalid:{resolved}")
-    payload = dict(cast(dict[str, Any], raw_payload))
-    schema_version = str(payload.get("schema_version") or "").strip()
-    if expected_schema_version and schema_version != expected_schema_version:
-        raise RuntimeError(
-            f"configured_empirical_payload_schema_mismatch:{resolved}:{schema_version}"
-        )
-    if not isinstance(payload.get("artifact_authority"), dict):
-        payload["artifact_authority"] = _empirical_artifact_authority(
-            evidence_name=evidence_name,
-            provenance=provenance,
-        )
-    return payload
-
-
-def _build_janus_q_summary_from_payloads(
-    *,
-    event_car_payload: dict[str, Any],
-    hgrm_reward_payload: dict[str, Any],
-    event_car_artifact_ref: str,
-    hgrm_reward_artifact_ref: str,
-) -> dict[str, object]:
-    event_summary = (
-        cast(dict[str, Any], event_car_payload.get("summary"))
-        if isinstance(event_car_payload.get("summary"), dict)
-        else {}
-    )
-    reward_summary = (
-        cast(dict[str, Any], hgrm_reward_payload.get("summary"))
-        if isinstance(hgrm_reward_payload.get("summary"), dict)
-        else {}
-    )
-    event_count = int(event_summary.get("event_count", 0) or 0)
-    reward_count = int(reward_summary.get("reward_count", 0) or 0)
-    mapped_count = int(reward_summary.get("event_mapped_count", 0) or 0)
-    reasons: list[str] = []
-    if event_count <= 0:
-        reasons.append("janus_event_count_missing")
-    if reward_count <= 0:
-        reasons.append("janus_reward_count_missing")
-    if reward_count > 0 and mapped_count < reward_count:
-        reasons.append("janus_reward_event_mapping_incomplete")
-    return {
-        "schema_version": "janus-q-evidence-v1",
-        "evidence_complete": not reasons,
-        "reasons": reasons,
-        "event_car": {
-            "schema_version": event_car_payload.get("schema_version"),
-            "event_count": event_count,
-            "manifest_hash": event_car_payload.get("manifest_hash"),
-            "artifact_ref": event_car_artifact_ref,
-        },
-        "hgrm_reward": {
-            "schema_version": hgrm_reward_payload.get("schema_version"),
-            "reward_count": reward_count,
-            "event_mapped_count": mapped_count,
-            "direction_gate_pass_ratio": str(
-                reward_summary.get("direction_gate_pass_ratio", "0")
-            ),
-            "manifest_hash": hgrm_reward_payload.get("manifest_hash"),
-            "artifact_ref": hgrm_reward_artifact_ref,
-        },
-    }
-
-
-def _build_bridge_evidence_payload(
-    report_payload: dict[str, Any],
-    *,
-    artifact_ref: str,
-    summary_fields: Sequence[str],
-) -> dict[str, Any]:
-    artifact_authority = report_payload.get("artifact_authority")
-    payload: dict[str, Any] = {
-        "artifact_ref": artifact_ref,
-        "artifact_authority": (
-            cast(dict[str, Any], artifact_authority)
-            if isinstance(artifact_authority, dict)
-            else {}
-        ),
-        "schema_version": str(report_payload.get("schema_version", "")).strip(),
-        "status": str(report_payload.get("status", "")).strip(),
-    }
-    for field_name in summary_fields:
-        if field_name in report_payload:
-            payload[field_name] = report_payload[field_name]
-    return payload
-
-
-def _build_vnext_gate_summary(
-    *,
-    runtime_strategies: list[StrategyRuntimeConfig],
-    simulation_calibration_payload: dict[str, Any],
-    shadow_live_deviation_payload: dict[str, Any],
-    dependency_quorum_payload: Mapping[str, Any] | None = None,
-    hypothesis_registry_payload: Mapping[str, Any] | None = None,
-) -> dict[str, Any]:
-    payload: dict[str, Any] = {
-        "strategy_compilation": [
-            {
-                "strategy_id": strategy.strategy_id,
-                "strategy_type": strategy.strategy_type,
-                "version": strategy.version,
-                "compiler_source": strategy.compiler_source,
-                "spec_compiled": strategy.compiler_source == "spec_v2",
-            }
-            for strategy in runtime_strategies
-        ],
-        "simulation_calibration": {
-            key: simulation_calibration_payload.get(key)
-            for key in (
-                "schema_version",
-                "status",
-                "order_count",
-                "expected_shortfall_sample_count",
-                "expected_shortfall_coverage",
-                "avg_calibration_error_bps",
-                "confidence_gate_action",
-                "artifact_authority",
-            )
-            if key in simulation_calibration_payload
-        },
-        "shadow_live_deviation": {
-            key: shadow_live_deviation_payload.get(key)
-            for key in (
-                "schema_version",
-                "status",
-                "order_count",
-                "decision_count",
-                "trade_count",
-                "avg_abs_slippage_bps",
-                "avg_abs_divergence_bps",
-                "deviation_budget_utilization",
-                "artifact_authority",
-            )
-            if key in shadow_live_deviation_payload
-        },
-        "portfolio_promotion": _build_portfolio_promotion_summary(runtime_strategies),
-    }
-    if dependency_quorum_payload is not None:
-        payload["dependency_quorum"] = dict(dependency_quorum_payload)
-    if hypothesis_registry_payload is not None:
-        payload["hypothesis_registry"] = dict(hypothesis_registry_payload)
-    return payload
-
-
-def _build_portfolio_promotion_summary(
-    runtime_strategies: Sequence[StrategyRuntimeConfig],
-) -> dict[str, Any]:
-    strategy_items = list(runtime_strategies)
-    spec_compiled_count = sum(
-        1 for item in strategy_items if item.compiler_source == "spec_v2"
-    )
-    symbol_usage: dict[str, list[str]] = {}
-    missing_policy_refs: set[str] = set()
-    promotion_policy_refs: set[str] = set()
-    risk_profile_refs: set[str] = set()
-    sizing_policy_refs: set[str] = set()
-    execution_policy_refs: set[str] = set()
-
-    for strategy in strategy_items:
-        universe = (
-            cast(dict[str, Any], strategy.strategy_spec.get("universe"))
-            if isinstance(strategy.strategy_spec.get("universe"), dict)
-            else {}
-        )
-        raw_symbols = universe.get("symbols")
-        if isinstance(raw_symbols, list):
-            for raw_symbol in cast(list[object], raw_symbols):
-                symbol = str(raw_symbol).strip().upper()
-                if not symbol:
-                    continue
-                symbol_usage.setdefault(symbol, []).append(strategy.strategy_id)
-        promotion_metadata = (
-            cast(dict[str, Any], strategy.compiled_targets.get("promotion_metadata"))
-            if isinstance(strategy.compiled_targets.get("promotion_metadata"), dict)
-            else {}
-        )
-        promotion_policy_ref = str(
-            promotion_metadata.get("promotion_policy_ref") or ""
-        ).strip()
-        risk_profile_ref = str(promotion_metadata.get("risk_profile_ref") or "").strip()
-        sizing_policy_ref = str(
-            promotion_metadata.get("sizing_policy_ref") or ""
-        ).strip()
-        execution_policy_ref = str(
-            promotion_metadata.get("execution_policy_ref") or ""
-        ).strip()
-        if promotion_policy_ref:
-            promotion_policy_refs.add(promotion_policy_ref)
-        else:
-            missing_policy_refs.add(f"{strategy.strategy_id}:promotion_policy_ref")
-        if risk_profile_ref:
-            risk_profile_refs.add(risk_profile_ref)
-        else:
-            missing_policy_refs.add(f"{strategy.strategy_id}:risk_profile_ref")
-        if sizing_policy_ref:
-            sizing_policy_refs.add(sizing_policy_ref)
-        else:
-            missing_policy_refs.add(f"{strategy.strategy_id}:sizing_policy_ref")
-        if execution_policy_ref:
-            execution_policy_refs.add(execution_policy_ref)
-        else:
-            missing_policy_refs.add(f"{strategy.strategy_id}:execution_policy_ref")
-
-    overlapping_symbols = sorted(
-        symbol for symbol, owners in symbol_usage.items() if len(set(owners)) > 1
-    )
-    return {
-        "mode": "portfolio_aware" if len(strategy_items) > 1 else "single_strategy",
-        "strategy_count": len(strategy_items),
-        "spec_compiled_count": spec_compiled_count,
-        "unique_symbol_count": len(symbol_usage),
-        "overlapping_symbols": overlapping_symbols,
-        "promotion_policy_refs": sorted(promotion_policy_refs),
-        "risk_profile_refs": sorted(risk_profile_refs),
-        "sizing_policy_refs": sorted(sizing_policy_refs),
-        "execution_policy_refs": sorted(execution_policy_refs),
-        "missing_policy_refs": sorted(missing_policy_refs),
-    }
-
-
-def _manifest_artifact_payload(
-    artifact_root: Path,
-    artifact_path: Path | None,
-    stage_name: str,
-    check_name: str,
-) -> dict[str, str] | None:
-    if artifact_path is None or not artifact_path.exists():
-        return None
-    artifact_payload = _load_json_if_exists(artifact_path)
-    contract = contract_from_artifact_payload(artifact_payload)
-    if not contract:
-        default_contract = _artifact_authority_for_check(
-            stage_name=stage_name,
-            check_name=check_name,
-        )
-        if default_contract:
-            contract = default_contract
-    payload: dict[str, Any] = {
-        "path": _manifest_relative_path(artifact_root, artifact_path),
-        "sha256": _sha256_path(artifact_path),
-        "stage": stage_name,
-        "check": check_name,
-    }
-    if contract:
-        payload["artifact_authority"] = contract
-    return cast(dict[str, str], payload)
-
-
-def _load_json_if_exists(path: Path) -> dict[str, Any] | None:
-    if not path.exists():
-        return None
-    try:
-        payload = json.loads(path.read_text(encoding="utf-8"))
-    except (OSError, json.JSONDecodeError):
-        return None
-    if not isinstance(payload, dict):
-        return None
-    return cast(dict[str, Any], payload)
-
-
-def _build_contamination_registry_payload(
-    *,
-    output_dir: Path,
-    run_id: str,
-    candidate_id: str,
-    now: datetime,
-    artifact_refs: Sequence[Path],
-) -> dict[str, Any]:
-    relative_refs = [
-        _manifest_relative_path(output_dir, artifact_path)
-        for artifact_path in artifact_refs
-    ]
-    payload: dict[str, Any] = {
-        "schema_version": "contamination-leakage-report-v1",
-        "run_id": run_id,
-        "candidate_id": candidate_id,
-        "generated_at": now.isoformat(),
-        "status": "pass",
-        "leakage_detected": False,
-        "leakage_rate": 0.0,
-        "temporal_integrity": {
-            "event_time_ordering_passed": True,
-            "embargo_windows_enforced": True,
-        },
-        "source_lineage": {
-            "complete": True,
-            "feature_sources": relative_refs,
-            "prompt_sources": [],
-        },
-        "checks": [
-            {"check": "temporal_ordering", "status": "pass"},
-            {"check": "lineage_complete", "status": "pass"},
-            {"check": "leakage_absent", "status": "pass"},
-            {"check": "embargo_windows_enforced", "status": "pass"},
-        ],
-        "artifact_refs": relative_refs,
-        "artifact_authority": evidence_contract_payload(
-            provenance=ArtifactProvenance.HISTORICAL_MARKET_REPLAY,
-            maturity=EvidenceMaturity.UNCALIBRATED,
-            calibration_summary={"status": "pending_calibration"},
-        ),
-    }
-    payload["artifact_hash"] = _stable_hash(
-        {key: value for key, value in payload.items() if key != "artifact_hash"}
-    )
-    return payload
-
-
-def _normalize_hmm_regime_id(value: object) -> str:
-    normalized = str(value).strip().lower()
-    return normalized or HMM_UNKNOWN_REGIME_ID
-
-
-def _build_hmm_state_posterior_payload(
-    *,
-    output_dir: Path,
-    run_id: str,
-    candidate_id: str,
-    now: datetime,
-    walk_decisions: Sequence[WalkForwardDecision],
-    walkforward_results_path: Path,
-    gate_policy_path: Path,
-) -> dict[str, Any]:
-    regime_counts: dict[str, int] = {}
-    entropy_band_counts: dict[str, int] = {}
-    guardrail_reason_counts: dict[str, int] = {}
-    posterior_mass: dict[str, Decimal] = {}
-    authoritative_samples = 0
-    transition_shock_samples = 0
-    stale_or_defensive_samples = 0
-
-    decision_timestamps = sorted(
-        {
-            item.decision.event_ts.isoformat()
-            for item in walk_decisions
-            if item.decision.event_ts.tzinfo is not None
-        }
-    )
-
-    for decision in walk_decisions:
-        context = resolve_hmm_context(decision.decision.params)
-        regime_id = _normalize_hmm_regime_id(context.regime_id)
-        entropy_band = str(context.entropy_band).strip().lower() or "unknown"
-        guardrail_reason = str(context.guardrail_reason or "").strip() or "none"
-
-        regime_counts[regime_id] = regime_counts.get(regime_id, 0) + 1
-        entropy_band_counts[entropy_band] = entropy_band_counts.get(entropy_band, 0) + 1
-        guardrail_reason_counts[guardrail_reason] = (
-            guardrail_reason_counts.get(guardrail_reason, 0) + 1
-        )
-
-        if context.is_authoritative:
-            authoritative_samples += 1
-        if context.transition_shock:
-            transition_shock_samples += 1
-        if context.guardrail.stale or context.guardrail.fallback_to_defensive:
-            stale_or_defensive_samples += 1
-
-        for posterior_regime, posterior_value in context.posterior.items():
-            try:
-                posterior_decimal = Decimal(str(posterior_value))
-            except Exception:
-                continue
-            if posterior_decimal.is_nan() or posterior_decimal.is_infinite():
-                continue
-            normalized_regime = _normalize_hmm_regime_id(posterior_regime)
-            posterior_mass[normalized_regime] = (
-                posterior_mass.get(normalized_regime, Decimal("0")) + posterior_decimal
-            )
-
-    sample_count = len(walk_decisions)
-    authoritative_ratio = (
-        Decimal(authoritative_samples) / Decimal(sample_count)
-        if sample_count > 0
-        else Decimal("0")
-    )
-    ordered_posterior_mass = {
-        regime: str(value)
-        for regime, value in sorted(
-            posterior_mass.items(),
-            key=lambda item: (-item[1], item[0]),
-        )
-    }
-    top_regime = (
-        max(posterior_mass.items(), key=lambda item: item[1])[0]
-        if posterior_mass
-        else HMM_UNKNOWN_REGIME_ID
-    )
-
-    payload: dict[str, Any] = {
-        "schema_version": "hmm-state-posterior-v1",
-        "run_id": run_id,
-        "candidate_id": candidate_id,
-        "generated_at": now.isoformat(),
-        "samples_total": sample_count,
-        "authoritative_samples": authoritative_samples,
-        "authoritative_sample_ratio": str(authoritative_ratio),
-        "transition_shock_samples": transition_shock_samples,
-        "stale_or_defensive_samples": stale_or_defensive_samples,
-        "regime_counts": dict(sorted(regime_counts.items())),
-        "entropy_band_counts": dict(sorted(entropy_band_counts.items())),
-        "guardrail_reason_counts": dict(sorted(guardrail_reason_counts.items())),
-        "posterior_mass_by_regime": ordered_posterior_mass,
-        "top_regime_by_posterior_mass": top_regime,
-        "decision_window": {
-            "first_event_ts": decision_timestamps[0] if decision_timestamps else None,
-            "last_event_ts": decision_timestamps[-1] if decision_timestamps else None,
-        },
-        "source_lineage": {
-            "walkforward_results_artifact_ref": _manifest_relative_path(
-                output_dir, walkforward_results_path
-            ),
-            "gate_policy_artifact_ref": _manifest_relative_path(
-                output_dir, gate_policy_path
-            ),
-            "decision_source": "walkforward_results",
-        },
-        "artifact_authority": evidence_contract_payload(
-            provenance=ArtifactProvenance.SYNTHETIC_GENERATED,
-            maturity=EvidenceMaturity.STUB,
-            authoritative=False,
-            placeholder=True,
-            notes="Current HMM posterior is a schema/contract scaffold, not trained empirical output.",
-        ),
-    }
-    payload["artifact_hash"] = _stable_hash(
-        {key: value for key, value in payload.items() if key != "artifact_hash"}
-    )
-    return payload
-
-
-def _decimal_or_zero(value: object) -> Decimal:
-    try:
-        decimal_value = Decimal(str(value))
-    except Exception:
-        return Decimal("0")
-    if decimal_value.is_nan() or decimal_value.is_infinite():
-        return Decimal("0")
-    return decimal_value
-
-
-def _decimal_or_none(value: object) -> Decimal | None:
-    try:
-        decimal_value = Decimal(str(value))
-    except Exception:
-        return None
-    if decimal_value.is_nan() or decimal_value.is_infinite():
-        return None
-    return decimal_value
-
-
-def _normalize_expert_weights(weights: Mapping[str, Decimal]) -> dict[str, Decimal]:
-    total = Decimal("0")
-    for value in weights.values():
-        if value > 0:
-            total += value
-    if total <= 0:
-        return {
-            "trend": Decimal("0.05"),
-            "reversal": Decimal("0.05"),
-            "breakout": Decimal("0.10"),
-            "defensive": Decimal("0.80"),
-        }
-    return {
-        "trend": max(Decimal("0"), weights.get("trend", Decimal("0"))) / total,
-        "reversal": max(Decimal("0"), weights.get("reversal", Decimal("0"))) / total,
-        "breakout": max(Decimal("0"), weights.get("breakout", Decimal("0"))) / total,
-        "defensive": max(Decimal("0"), weights.get("defensive", Decimal("0"))) / total,
-    }
-
-
-def _build_expert_router_weights(
-    *,
-    regime_label: str,
-    context: Any,
-) -> tuple[dict[str, Decimal], bool]:
-    fallback_active = bool(
-        context.guardrail.fallback_to_defensive
-        or context.guardrail.stale
-        or context.transition_shock
-    )
-    if context.entropy_band == "high":
-        fallback_active = True
-    if fallback_active:
-        return (
-            _normalize_expert_weights(
-                {
-                    "trend": Decimal("0.05"),
-                    "reversal": Decimal("0.05"),
-                    "breakout": Decimal("0.10"),
-                    "defensive": Decimal("0.80"),
-                }
-            ),
-            True,
-        )
-
-    normalized_regime_label = regime_label.strip().lower()
-    if normalized_regime_label in {"r2", "trend", "trend_up"}:
-        return (
-            _normalize_expert_weights(
-                {
-                    "trend": Decimal("0.62"),
-                    "reversal": Decimal("0.14"),
-                    "breakout": Decimal("0.20"),
-                    "defensive": Decimal("0.04"),
-                }
-            ),
-            False,
-        )
-    if normalized_regime_label in {"r3", "breakout", "risk_on_breakout"}:
-        return (
-            _normalize_expert_weights(
-                {
-                    "trend": Decimal("0.24"),
-                    "reversal": Decimal("0.11"),
-                    "breakout": Decimal("0.55"),
-                    "defensive": Decimal("0.10"),
-                }
-            ),
-            False,
-        )
-    if normalized_regime_label in {"r1", "range", "mean_revert", "reversal"}:
-        return (
-            _normalize_expert_weights(
-                {
-                    "trend": Decimal("0.18"),
-                    "reversal": Decimal("0.55"),
-                    "breakout": Decimal("0.12"),
-                    "defensive": Decimal("0.15"),
-                }
-            ),
-            False,
-        )
-    if normalized_regime_label in {"r4", "r5", "stressed", "stress"}:
-        return (
-            _normalize_expert_weights(
-                {
-                    "trend": Decimal("0.08"),
-                    "reversal": Decimal("0.12"),
-                    "breakout": Decimal("0.10"),
-                    "defensive": Decimal("0.70"),
-                }
-            ),
-            False,
-        )
-
-    return (
-        _normalize_expert_weights(
-            {
-                "trend": Decimal("0.28"),
-                "reversal": Decimal("0.22"),
-                "breakout": Decimal("0.26"),
-                "defensive": Decimal("0.24"),
-            }
-        ),
-        False,
-    )
-
-
-def _build_expert_router_registry_payload(
-    *,
-    output_dir: Path,
-    run_id: str,
-    candidate_id: str,
-    now: datetime,
-    walk_decisions: Sequence[WalkForwardDecision],
-    walkforward_results_path: Path,
-    gate_policy_path: Path,
-    strategy_config_path: Path,
-    hmm_state_posterior_path: Path,
-    policy_payload: Mapping[str, Any],
-) -> dict[str, Any]:
-    expert_sums: dict[str, Decimal] = {
-        "trend": Decimal("0"),
-        "reversal": Decimal("0"),
-        "breakout": Decimal("0"),
-        "defensive": Decimal("0"),
-    }
-    top_expert_counts: dict[str, int] = {
-        "trend": 0,
-        "reversal": 0,
-        "breakout": 0,
-        "defensive": 0,
-    }
-    fallback_count = 0
-    max_expert_weight = Decimal("0")
-
-    for decision in walk_decisions:
-        params = decision.decision.params
-        context = resolve_hmm_context(params)
-        route_regime_label = str(
-            params.get("route_regime_label")
-            or params.get("regime_label")
-            or context.regime_id
-        )
-        weights, fallback_active = _build_expert_router_weights(
-            regime_label=route_regime_label,
-            context=context,
-        )
-        if fallback_active:
-            fallback_count += 1
-
-        top_expert = "defensive"
-        top_weight = Decimal("-1")
-        for expert_name in ("trend", "reversal", "breakout", "defensive"):
-            weight = weights.get(expert_name, Decimal("0"))
-            expert_sums[expert_name] = expert_sums[expert_name] + weight
-            if weight > top_weight:
-                top_weight = weight
-                top_expert = expert_name
-        top_expert_counts[top_expert] = top_expert_counts.get(top_expert, 0) + 1
-        if top_weight > max_expert_weight:
-            max_expert_weight = top_weight
-
-    route_count = len(walk_decisions)
-    if route_count > 0:
-        fallback_rate = Decimal(fallback_count) / Decimal(route_count)
-    else:
-        fallback_rate = Decimal("0")
-    if route_count > 0:
-        avg_expert_weights = {
-            expert: str(expert_sums[expert] / Decimal(route_count))
-            for expert in ("trend", "reversal", "breakout", "defensive")
-        }
-    else:
-        avg_expert_weights = {
-            "trend": "0",
-            "reversal": "0",
-            "breakout": "0",
-            "defensive": "0",
-        }
-
-    max_fallback_rate = _decimal_or_zero(
-        policy_payload.get("promotion_expert_router_max_fallback_rate")
-    )
-    if max_fallback_rate <= 0:
-        max_fallback_rate = Decimal("0.05")
-    max_concentration = _decimal_or_zero(
-        policy_payload.get("promotion_expert_router_max_expert_concentration")
-    )
-    if max_concentration <= 0:
-        max_concentration = Decimal("0.85")
-
-    fallback_slo_pass = fallback_rate <= max_fallback_rate
-    concentration_slo_pass = max_expert_weight <= max_concentration
-    reasons: list[str] = []
-    if not fallback_slo_pass:
-        reasons.append("fallback_rate_exceeds_threshold")
-    if not concentration_slo_pass:
-        reasons.append("expert_concentration_exceeds_threshold")
-    if route_count <= 0:
-        reasons.append("router_decisions_missing")
-
-    dominant_expert = "defensive"
-    dominant_expert_count = -1
-    for expert_name in ("trend", "reversal", "breakout", "defensive"):
-        count = top_expert_counts.get(expert_name, 0)
-        if count > dominant_expert_count:
-            dominant_expert = expert_name
-            dominant_expert_count = count
-
-    payload: dict[str, Any] = {
-        "schema_version": "expert-router-registry-v1",
-        "run_id": run_id,
-        "candidate_id": candidate_id,
-        "generated_at": now.isoformat(),
-        "router_version": "router-v1",
-        "route_count": route_count,
-        "fallback_count": fallback_count,
-        "fallback_rate": str(fallback_rate),
-        "max_expert_weight": str(max_expert_weight),
-        "avg_expert_weights": avg_expert_weights,
-        "top_expert_counts": dict(sorted(top_expert_counts.items())),
-        "concentration": {
-            "dominant_expert": dominant_expert,
-            "dominant_expert_count": dominant_expert_count,
-            "max_expert_weight": str(max_expert_weight),
-        },
-        "slo_feedback": {
-            "max_fallback_rate": str(max_fallback_rate),
-            "max_expert_concentration": str(max_concentration),
-            "fallback_rate": str(fallback_rate),
-            "max_observed_expert_weight": str(max_expert_weight),
-            "fallback_slo_pass": fallback_slo_pass,
-            "concentration_slo_pass": concentration_slo_pass,
-            "overall_status": (
-                "pass"
-                if fallback_slo_pass and concentration_slo_pass and route_count > 0
-                else "fail"
-            ),
-            "reasons": reasons,
-        },
-        "source_lineage": {
-            "walkforward_results_artifact_ref": _manifest_relative_path(
-                output_dir, walkforward_results_path
-            ),
-            "hmm_state_posterior_artifact_ref": _manifest_relative_path(
-                output_dir, hmm_state_posterior_path
-            ),
-            "gate_policy_artifact_ref": _manifest_relative_path(
-                output_dir, gate_policy_path
-            ),
-            "strategy_config_artifact_ref": _manifest_relative_path(
-                output_dir, strategy_config_path
-            ),
-        },
-        "artifact_authority": evidence_contract_payload(
-            provenance=ArtifactProvenance.SYNTHETIC_GENERATED,
-            maturity=EvidenceMaturity.STUB,
-            authoritative=False,
-            placeholder=True,
-            notes="Router registry is currently derived from deterministic strategy scaffolding.",
-        ),
-    }
-    payload["artifact_hash"] = _stable_hash(
-        {key: value for key, value in payload.items() if key != "artifact_hash"}
-    )
-    return payload
-
-
-def _build_profitability_stage_manifest(
-    *,
-    output_dir: Path,
-    run_id: str,
-    candidate_id: str,
-    strategy_family: str,
-    llm_artifact_ref: str | None,
-    router_artifact_ref: str,
-    run_context: dict[str, str],
-    research_manifest_path: Path | None,
-    candidate_spec_path: Path,
-    evaluation_report_path: Path,
-    walkforward_results_path: Path,
-    baseline_evaluation_report_path: Path,
-    gate_report_payload: dict[str, Any],
-    gate_report_path: Path,
-    profitability_benchmark_path: Path,
-    contamination_registry_path: Path,
-    benchmark_parity_path: Path,
-    foundation_router_parity_path: Path,
-    deeplob_bdlob_report_path: Path,
-    advisor_fallback_slo_report_path: Path,
-    profitability_evidence_path: Path,
-    profitability_validation_path: Path,
-    simulation_calibration_report_path: Path,
-    shadow_live_deviation_report_path: Path,
-    hmm_state_posterior_path: Path,
-    expert_router_registry_path: Path,
-    janus_event_car_path: Path,
-    janus_hgrm_reward_path: Path,
-    recalibration_report_path: Path,
-    rollback_check: RollbackReadinessResult,
-    drift_gate_check: dict[str, Any],
-    patch_path: Path | None,
-    now: datetime,
-) -> dict[str, Any]:
-    research_artifacts: dict[str, dict[str, str]] = {}
-    research_checks: list[dict[str, object]] = []
-    candidate_spec_artifact = _manifest_artifact_payload(
-        output_dir,
-        candidate_spec_path,
-        "research",
-        "candidate_spec_present",
-    )
-    if candidate_spec_artifact:
-        research_artifacts["candidate_spec"] = candidate_spec_artifact
-        research_checks.append({"check": "candidate_spec_present", "status": "pass"})
-    else:
-        research_checks.append(
-            {"check": "candidate_spec_present", "status": "fail", "reason": "missing"}
-        )
-
-    if research_manifest_path is not None:
-        candidate_generation_manifest_artifact = _manifest_artifact_payload(
-            output_dir,
-            research_manifest_path,
-            "research",
-            "candidate_generation_manifest_present",
-        )
-        if candidate_generation_manifest_artifact:
-            research_artifacts["candidate_generation_manifest"] = (
-                candidate_generation_manifest_artifact
-            )
-            research_checks.append(
-                {"check": "candidate_generation_manifest_present", "status": "pass"}
-            )
-        else:
-            research_checks.append(
-                {
-                    "check": "candidate_generation_manifest_present",
-                    "status": "fail",
-                    "reason": "missing",
-                }
-            )
-
-    for artifact_path, check_name in (
-        (walkforward_results_path, "walkforward_results_present"),
-        (baseline_evaluation_report_path, "baseline_evaluation_report_present"),
-    ):
-        entry = _manifest_artifact_payload(
-            output_dir,
-            artifact_path,
-            "research",
-            check_name,
-        )
-        if entry:
-            research_artifacts[check_name] = entry
-            research_checks.append({"check": check_name, "status": "pass"})
-        else:
-            research_checks.append(
-                {"check": check_name, "status": "fail", "reason": "missing"}
-            )
-
-    research_pass = all(item.get("status") == "pass" for item in research_checks)
-
-    validation_artifacts: dict[str, dict[str, str]] = {}
-    validation_checks: list[dict[str, object]] = []
-    for artifact_path, check_name in (
-        (evaluation_report_path, "evaluation_report_present"),
-        (contamination_registry_path, "contamination_registry_present"),
-        (profitability_benchmark_path, "profitability_benchmark_present"),
-        (profitability_evidence_path, "profitability_evidence_present"),
-        (profitability_validation_path, "profitability_validation_present"),
-        (
-            simulation_calibration_report_path,
-            "simulation_calibration_report_present",
-        ),
-        (
-            shadow_live_deviation_report_path,
-            "shadow_live_deviation_report_present",
-        ),
-    ):
-        entry = _manifest_artifact_payload(
-            output_dir,
-            artifact_path,
-            "validation",
-            check_name,
-        )
-        if entry:
-            validation_artifacts[check_name] = entry
-            validation_checks.append({"check": check_name, "status": "pass"})
-        else:
-            validation_checks.append(
-                {"check": check_name, "status": "fail", "reason": "missing"}
-            )
-
-    validation_pass = all(item.get("status") == "pass" for item in validation_checks)
-    if (
-        profitability_validation_path.exists()
-        and _load_json_if_exists(profitability_validation_path) is None
-    ):
-        validation_checks.append(
-            {
-                "check": "profitability_validation_payload_valid_json",
-                "status": "fail",
-                "reason": "invalid_json",
-            }
-        )
-        validation_pass = False
-
-    execution_artifacts: dict[str, dict[str, str]] = {}
-    execution_checks_map: list[tuple[str, str, Path]] = [
-        (
-            "walkforward_results_present",
-            "walkforward_results",
-            walkforward_results_path,
-        ),
-        ("evaluation_report_present", "evaluation_report", evaluation_report_path),
-        ("gate_evaluation_present", "gate_evaluation", gate_report_path),
-        ("benchmark_parity_present", "benchmark_parity", benchmark_parity_path),
-        (
-            "foundation_router_parity_present",
-            "foundation_router_parity",
-            foundation_router_parity_path,
-        ),
-        (
-            "deeplob_bdlob_contract_present",
-            "deeplob_bdlob_contract",
-            deeplob_bdlob_report_path,
-        ),
-        (
-            "advisor_fallback_slo_present",
-            "advisor_fallback_slo",
-            advisor_fallback_slo_report_path,
-        ),
-        (
-            "simulation_calibration_report_present",
-            "simulation_calibration",
-            simulation_calibration_report_path,
-        ),
-        (
-            "shadow_live_deviation_report_present",
-            "shadow_live_deviation",
-            shadow_live_deviation_report_path,
-        ),
-        (
-            "hmm_state_posterior_present",
-            "hmm_state_posterior",
-            hmm_state_posterior_path,
-        ),
-        (
-            "expert_router_registry_present",
-            "expert_router_registry",
-            expert_router_registry_path,
-        ),
-        ("janus_event_car_present", "janus_event_car", janus_event_car_path),
-        ("janus_hgrm_reward_present", "janus_hgrm_reward", janus_hgrm_reward_path),
-        (
-            "recalibration_report_present",
-            "recalibration_report",
-            recalibration_report_path,
-        ),
-    ]
-    for check_name, artifact_key, artifact_path in execution_checks_map:
-        payload = (
-            _manifest_artifact_payload(
-                output_dir,
-                artifact_path,
-                "execution",
-                check_name,
-            )
-            or {}
-        )
-        if payload:
-            execution_artifacts[artifact_key] = payload
-    execution_checks: list[dict[str, object]] = []
-    for check_name, artifact_key, _artifact_path in execution_checks_map:
-        payload = execution_artifacts.get(artifact_key, {})
-        status = "pass" if payload else "fail"
-        execution_checks.append({"check": check_name, "status": status})
-        if status != "pass":
-            execution_checks[-1]["reason"] = "missing"
-    gate_allowance = bool(gate_report_payload.get("promotion_allowed", False))
-    drift_allowed = bool(drift_gate_check.get("allowed", False))
-    execution_checks.extend(
-        [
-            {
-                "check": "gate_matrix_approval",
-                "status": "pass" if gate_allowance else "fail",
-            },
-            {
-                "check": "drift_gate_approval",
-                "status": "pass" if drift_allowed else "fail",
-            },
-        ]
-    )
-    execution_pass = all(item["status"] == "pass" for item in execution_checks)
-
-    governance_artifacts: dict[str, dict[str, str]] = {
-        "candidate_spec": (
-            _manifest_artifact_payload(
-                output_dir,
-                candidate_spec_path,
-                "governance",
-                "candidate_spec_present",
-            )
-            or {}
-        ),
-        "gate_evaluation": (
-            _manifest_artifact_payload(
-                output_dir,
-                gate_report_path,
-                "governance",
-                "gate_evaluation_present",
-            )
-            or {}
-        ),
-        "rollback_readiness": (
-            _manifest_artifact_payload(
-                output_dir,
-                output_dir / "gates" / "rollback-readiness.json",
-                "governance",
-                "rollback_readiness_present",
-            )
-            or {}
-        ),
-    }
-    if patch_path is not None:
-        governance_artifacts["paper_patch"] = (
-            _manifest_artifact_payload(
-                output_dir,
-                patch_path,
-                "governance",
-                "paper_patch_present",
-            )
-            or {}
-        )
-    governance_checks = [
-        {
-            "check": "rollback_ready",
-            "status": "pass" if rollback_check.ready else "fail",
-        },
-        {
-            "check": "rollback_readiness_present",
-            "status": "pass"
-            if bool(governance_artifacts.get("rollback_readiness")) or False
-            else "fail",
-        },
-        {
-            "check": "gate_report_present",
-            "status": "pass" if gate_report_path.exists() else "fail",
-        },
-    ]
-    if patch_path is None:
-        governance_checks.append(
-            {
-                "check": "paper_patch_present",
-                "status": "pass",
-            }
-        )
-    governance_checks.extend(
-        [
-            {
-                "check": "candidate_spec_present",
-                "status": "pass" if candidate_spec_path.exists() else "fail",
-            },
-            {"check": "risk_controls_attestable", "status": "pass"},
-        ]
-    )
-    governance_pass = all(item["status"] == "pass" for item in governance_checks)
-
-    stage_payloads: dict[str, dict[str, Any]] = {
-        "research": {
-            "status": "pass" if research_pass else "fail",
-            "checks": research_checks,
-            "artifacts": research_artifacts,
-            "owner": "research-orchestrator",
-            "completed_at_utc": now.isoformat(),
-        },
-        "validation": {
-            "status": "pass" if validation_pass else "fail",
-            "checks": validation_checks,
-            "artifacts": validation_artifacts,
-            "owner": "validation-service",
-            "completed_at_utc": now.isoformat(),
-        },
-        "execution": {
-            "status": "pass" if execution_pass else "fail",
-            "checks": execution_checks,
-            "artifacts": {
-                key: value for key, value in execution_artifacts.items() if value
-            },
-            "owner": "execution-sim",
-            "completed_at_utc": now.isoformat(),
-        },
-        "governance": {
-            "status": "pass" if governance_pass else "fail",
-            "checks": governance_checks,
-            "artifacts": {
-                key: value for key, value in governance_artifacts.items() if value
-            },
-            "owner": "governance-policy",
-            "completed_at_utc": now.isoformat(),
-        },
-    }
-    overall_status = (
-        "pass"
-        if all(item["status"] == "pass" for item in stage_payloads.values())
-        else "fail"
-    )
-    failure_reasons = [
-        item["check"]
-        for stage in stage_payloads.values()
-        for item in stage["checks"]
-        if item.get("status") == "fail"
-    ]
-    replay_contract_artifact_hashes: dict[str, str] = {}
-    for stage_payload in stage_payloads.values():
-        artifacts_raw = stage_payload.get("artifacts")
-        if not isinstance(artifacts_raw, dict):
-            continue
-        artifacts = cast(dict[str, object], artifacts_raw)
-        for artifact_payload_raw in artifacts.values():
-            if not isinstance(artifact_payload_raw, dict):
-                continue
-            artifact_payload = cast(dict[str, object], artifact_payload_raw)
-            artifact_ref = str(artifact_payload.get("path", "")).strip()
-            artifact_sha = str(artifact_payload.get("sha256", "")).strip()
-            if not artifact_ref or not artifact_sha:
-                continue
-            replay_contract_artifact_hashes[artifact_ref] = artifact_sha
-    payload: dict[str, Any] = {
-        "schema_version": _PROFITABILITY_STAGE_MANIFEST_SCHEMA_VERSION,
-        "candidate_id": candidate_id,
-        "strategy_family": strategy_family,
-        "llm_artifact_ref": llm_artifact_ref,
-        "router_artifact_ref": router_artifact_ref,
-        "hmm_state_posterior_artifact_ref": _manifest_relative_path(
-            output_dir, hmm_state_posterior_path
-        ),
-        "expert_router_registry_artifact_ref": _manifest_relative_path(
-            output_dir, expert_router_registry_path
-        ),
-        "run_context": {
-            "repository": run_context.get("repository", "unknown"),
-            "base": run_context.get("base", "main"),
-            "head": run_context.get("head", "unknown"),
-            "artifact_path": run_context.get("artifact_path", str(output_dir)),
-            "run_id": run_context.get("run_id", run_id),
-            "design_doc": run_context.get("design_doc", ""),
-            "priority_id": run_context.get("priority_id", ""),
-        },
-        "stages": stage_payloads,
-        "overall_status": overall_status,
-        "failure_reasons": failure_reasons,
-        "replay_contract": {
-            "artifact_hashes": replay_contract_artifact_hashes,
-            "contract_hash": _stable_hash(
-                {"artifact_hashes": replay_contract_artifact_hashes}
-            ),
-            "hash_algorithm": "sha256",
-        },
-        "rollback_contract_ref": _manifest_relative_path(
-            output_dir, output_dir / "gates" / "rollback-readiness.json"
-        ),
-        "created_at_utc": now.isoformat(),
-    }
-    manifest_copy = dict(payload)
-    payload["content_hash"] = _stable_hash(manifest_copy)
-    return payload
-
-
-def _write_iteration_notes(
-    *,
-    artifact_root: Path,
-    run_id: str,
-    candidate_id: str,
-    stage_records: Sequence[_StageManifestRecord],
-    repository: str | None,
-    base: str | None,
-    head: str | None,
-    priority_id: str | None,
-) -> Path:
-    notes_dir = artifact_root / "notes"
-    notes_dir.mkdir(parents=True, exist_ok=True)
-    iteration = _readable_notes_iteration_number(notes_dir)
-    notes_path = notes_dir / f"iteration-{iteration}.md"
-
-    lines = [
-        f"# Autonomous lane iteration {iteration}",
-        "",
-        f"- run_id: {run_id}",
-        f"- candidate_id: {candidate_id}",
-    ]
-    if repository:
-        lines.append(f"- repository: {repository}")
-    if base:
-        lines.append(f"- base: {base}")
-    if head:
-        lines.append(f"- head: {head}")
-    if priority_id:
-        lines.append(f"- priority_id: {priority_id}")
-
-    lines.extend(["", "## Stage progression", ""])
-    for record in stage_records:
-        lines.append(f"- {record.stage} (index {record.stage_index})")
-        lines.append(f"  - trace={record.stage_trace_id}")
-        if record.parent_stage:
-            lines.append(
-                f"  - parent: {record.parent_stage} (hash={record.parent_lineage_hash})"
-            )
-    notes_path.write_text("\n".join(lines), encoding="utf-8")
-    return notes_path
-
-
-def _extract_janus_q_metrics(
-    summary: dict[str, object],
-) -> tuple[int, int, bool, list[str]]:
-    event_car = _as_object_dict(summary.get("event_car"))
-    hgrm_reward = _as_object_dict(summary.get("hgrm_reward"))
-    reasons_raw = summary.get("reasons")
-    reasons: list[str] = []
-    if isinstance(reasons_raw, list):
-        reason_values = cast(list[object], reasons_raw)
-        reasons = [
-            str(reason).strip() for reason in reason_values if str(reason).strip()
-        ]
-    return (
-        _safe_int(event_car.get("event_count", 0)),
-        _safe_int(hgrm_reward.get("reward_count", 0)),
-        bool(summary.get("evidence_complete", False)),
-        reasons,
-    )
 
 
 def run_autonomous_lane(
@@ -4475,750 +2721,6 @@ def run_autonomous_lane(
             now=now,
         )
         raise RuntimeError(f"autonomous_lane_persistence_failed: {exc}") from exc
-
-
-def _prepare_lane_output_dirs(output_dir: Path) -> tuple[Path, Path, Path, Path, Path]:
-    output_dir.mkdir(parents=True, exist_ok=True)
-    research_dir = output_dir / "research"
-    backtest_dir = output_dir / "backtest"
-    gates_dir = output_dir / "gates"
-    paper_dir = output_dir / "paper-candidate"
-    rollout_dir = output_dir / "rollout"
-    for path in (research_dir, backtest_dir, gates_dir, paper_dir, rollout_dir):
-        path.mkdir(parents=True, exist_ok=True)
-    return research_dir, backtest_dir, gates_dir, paper_dir, rollout_dir
-
-
-def _normalize_governance_inputs(
-    governance_inputs: Mapping[str, Any] | None,
-) -> dict[str, Any]:
-    raw: Mapping[str, Any] = (
-        governance_inputs if isinstance(governance_inputs, Mapping) else {}
-    )
-    execution_context_raw = raw.get("execution_context", {})
-    execution_context: Mapping[str, Any] = (
-        cast(Mapping[str, Any], execution_context_raw)
-        if isinstance(execution_context_raw, Mapping)
-        else {}
-    )
-    runtime_raw = raw.get("runtime_governance", {})
-    runtime: Mapping[str, Any] = (
-        cast(Mapping[str, Any], runtime_raw) if isinstance(runtime_raw, Mapping) else {}
-    )
-    rollback_raw = raw.get("rollback_proof", {})
-    rollback: Mapping[str, Any] = (
-        cast(Mapping[str, Any], rollback_raw)
-        if isinstance(rollback_raw, Mapping)
-        else {}
-    )
-    return {
-        "execution_context": {
-            "repository": _coerce_str(
-                execution_context.get("repository"), default="unknown"
-            ),
-            "base": _coerce_str(execution_context.get("base"), default="unknown"),
-            "head": _coerce_str(execution_context.get("head"), default="unknown"),
-            "artifactPath": _coerce_str(
-                execution_context.get("artifactPath"),
-                default="",
-            ),
-            "priorityId": _coerce_str(
-                execution_context.get("priorityId"),
-                default="",
-            ),
-            "designDoc": _coerce_str(
-                execution_context.get("designDoc"),
-                default="",
-            ),
-        },
-        "runtime_governance": cast(Mapping[str, Any], runtime),
-        "rollback_proof": cast(Mapping[str, Any], rollback),
-    }
-
-
-def _coalesce_governance_context(
-    *,
-    governance_inputs: Mapping[str, Any] | None,
-    governance_repository: str | None,
-    governance_base: str | None,
-    governance_head: str | None,
-    governance_artifact_path: str | None,
-    design_doc: str | None,
-    priority_id: str | None,
-    now: datetime,
-) -> dict[str, Any]:
-    raw: Mapping[str, Any] = (
-        governance_inputs if isinstance(governance_inputs, Mapping) else {}
-    )
-    execution_context_raw = raw.get("execution_context", {})
-    execution_context: Mapping[str, Any] = (
-        cast(Mapping[str, Any], execution_context_raw)
-        if isinstance(execution_context_raw, Mapping)
-        else {}
-    )
-    runtime_raw = raw.get("runtime_governance", {})
-    runtime: Mapping[str, Any] = (
-        cast(Mapping[str, Any], runtime_raw) if isinstance(runtime_raw, Mapping) else {}
-    )
-    rollback_raw = raw.get("rollback_proof", {})
-    rollback: Mapping[str, Any] = (
-        cast(Mapping[str, Any], rollback_raw)
-        if isinstance(rollback_raw, Mapping)
-        else {}
-    )
-
-    fallback_head = _coerce_str(execution_context.get("head")) or (
-        f"agentruns/torghut-autonomy-{now.strftime('%Y%m%dT%H%M%S')}"
-    )
-    resolved_repository = _coerce_str(governance_repository)
-    if not resolved_repository:
-        resolved_repository = _coerce_str(execution_context.get("repository"))
-    if not resolved_repository:
-        resolved_repository = "proompteng/lab"
-    resolved_base = _coerce_str(governance_base)
-    if not resolved_base:
-        resolved_base = _coerce_str(execution_context.get("base"))
-    if not resolved_base:
-        resolved_base = "main"
-
-    return {
-        "execution_context": {
-            "repository": resolved_repository,
-            "base": resolved_base,
-            "head": _coerce_str(
-                governance_head or _coerce_str(fallback_head),
-                default="unknown",
-            ),
-            "artifactPath": (
-                _coerce_str(governance_artifact_path)
-                or _coerce_str(execution_context.get("artifactPath"))
-            ),
-            "priorityId": (
-                priority_id
-                if priority_id
-                else _coerce_str(execution_context.get("priorityId"))
-            ),
-            "designDoc": (
-                (
-                    design_doc
-                    if design_doc is not None
-                    else _coerce_str(execution_context.get("designDoc"))
-                )
-                or _V6_08_GOVERNING_DESIGN_DOC
-            ),
-        },
-        "runtime_governance": cast(Mapping[str, Any], runtime),
-        "rollback_proof": cast(Mapping[str, Any], rollback),
-    }
-
-
-def _build_phase_manifest(
-    *,
-    run_id: str,
-    candidate_id: str,
-    evaluated_at: datetime,
-    output_dir: Path,
-    signals: list[SignalEnvelope],
-    requested_promotion_target: PromotionTarget,
-    gate_report: GateEvaluationReport,
-    gate_report_payload: dict[str, Any],
-    gate_report_path: Path,
-    promotion_check: Any,
-    rollback_check: Any,
-    drift_gate_check: dict[str, Any],
-    patch_path: Path | None,
-    recommended_mode: str,
-    promotion_reasons: list[str],
-    governance_inputs: Mapping[str, Any] | None,
-    drift_promotion_evidence: dict[str, Any] | None,
-) -> dict[str, Any]:
-    phase_timestamp = evaluated_at.isoformat()
-    governance = _normalize_governance_inputs(governance_inputs)
-    execution_context = cast(Mapping[str, Any], governance["execution_context"])
-    artifact_path = _coerce_str(
-        execution_context.get("artifactPath"),
-        default=str(output_dir),
-    )
-    runtime_governance = cast(Mapping[str, Any], governance["runtime_governance"])
-    rollback_proof = cast(Mapping[str, Any], governance["rollback_proof"])
-    gate_status = "pass" if gate_report.promotion_allowed else "fail"
-    prerequisite_status = "pass" if promotion_check.allowed else "fail"
-    rollback_ready_status = "pass" if rollback_check.ready else "fail"
-    drift_gate_status = (
-        "pass" if bool(drift_gate_check.get("allowed", False)) else "fail"
-    )
-    canary_status = (
-        "pass"
-        if patch_path is not None
-        else ("fail" if requested_promotion_target == "paper" else "skipped")
-    )
-    canary_slo_status = (
-        "pass"
-        if patch_path is not None
-        else ("fail" if requested_promotion_target == "paper" else "skipped")
-    )
-    gate_payload_gates = gate_report_payload.get("gates")
-    gate_entries = (
-        cast(list[object], gate_payload_gates)
-        if isinstance(gate_payload_gates, list)
-        else []
-    )
-    slo_gates = _coerce_gate_phase_gates(gate_entries)
-    throughput_payload = cast(
-        Mapping[str, Any], gate_report_payload.get("throughput", {})
-    )
-    signal_count = _coerce_int(throughput_payload.get("signal_count"), default=0)
-    decision_count = _coerce_int(throughput_payload.get("decision_count"), default=0)
-    drift_refs_raw = drift_gate_check.get("artifact_refs")
-    drift_artifacts = _coerce_path_strings(
-        drift_refs_raw if isinstance(drift_refs_raw, list) else []
-    )
-    runtime_gate_status = coerce_phase_status(
-        runtime_governance.get("governance_status", "skipped")
-    )
-    runtime_refs_raw = runtime_governance.get("artifact_refs")
-    runtime_artifact_refs = _coerce_path_strings(
-        runtime_refs_raw if isinstance(runtime_refs_raw, list) else []
-    )
-    rollback_triggered = bool(
-        runtime_governance.get("rollback_triggered", False)
-        or rollback_proof.get("rollback_triggered", False)
-    )
-    rollback_proof_path = _coerce_str(
-        rollback_proof.get("rollback_incident_evidence_path"), default=""
-    )
-    if not rollback_proof_path:
-        rollback_proof_path = _coerce_str(
-            rollback_proof.get("rollback_incident_evidence"), default=""
-        )
-    rollback_proof_status = (
-        "pass"
-        if (not rollback_triggered)
-        else ("pass" if rollback_proof_path else "fail")
-    )
-    drift_evidence: Mapping[str, Any] = drift_promotion_evidence or {}
-    drift_evidence_refs_raw = drift_evidence.get("evidence_artifact_refs")
-    drift_evidence_refs = (
-        cast(list[object], drift_evidence_refs_raw)
-        if isinstance(drift_evidence_refs_raw, list)
-        else []
-    )
-    drift_gate_artifacts = sorted(
-        {str(item) for item in drift_evidence_refs if str(item).strip()}
-    )
-
-    phase_summaries: list[dict[str, Any]] = [
-        {
-            "name": "gate-evaluation",
-            "status": gate_status,
-            "timestamp": phase_timestamp,
-            "slo_gates": slo_gates
-            + [
-                {
-                    "id": "slo_signal_count_minimum",
-                    "status": "pass" if signal_count >= 1 else "fail",
-                    "threshold": 1,
-                    "value": signal_count,
-                },
-                {
-                    "id": "slo_decision_count_minimum",
-                    "status": "pass" if decision_count >= 1 else "fail",
-                    "threshold": 1,
-                    "value": decision_count,
-                },
-            ],
-            "observations": {
-                "recommended_mode": recommended_mode,
-                "promotion_reasons": promotion_reasons,
-                "throughput": throughput_payload,
-            },
-            "required": {
-                "signal_count": 1,
-                "decision_count": 1,
-                "trade_count": 0,
-            },
-            "artifact_refs": [
-                str(gate_report_path),
-                str(output_dir / "gates" / "promotion-evidence-gate.json"),
-            ],
-        },
-        {
-            "name": "promotion-prerequisites",
-            "status": prerequisite_status,
-            "timestamp": phase_timestamp,
-            "slo_gates": [
-                {
-                    "id": "slo_required_artifacts_present",
-                    "status": "pass"
-                    if not promotion_check.missing_artifacts
-                    else "fail",
-                    "threshold": len(promotion_check.required_artifacts),
-                    "value": len(promotion_check.artifact_refs),
-                }
-            ],
-            "observations": {
-                "artifact_requirements": list(promotion_check.required_artifacts),
-                "missing_artifacts": list(promotion_check.missing_artifacts),
-                "throughput_required": dict(promotion_check.required_throughput),
-                "throughput_observed": dict(promotion_check.observed_throughput),
-            },
-            "reasons": list(promotion_check.reasons),
-            "artifact_refs": [str(path) for path in promotion_check.artifact_refs],
-            "artifact_paths": {
-                "promotion_check": str(
-                    output_dir / "gates" / "promotion-prerequisites.json"
-                ),
-                "promotion_gate": str(
-                    output_dir / "gates" / "promotion-evidence-gate.json"
-                ),
-            },
-        },
-        {
-            "name": "rollback-readiness",
-            "status": rollback_ready_status,
-            "timestamp": phase_timestamp,
-            "slo_gates": [
-                {
-                    "id": "slo_required_rollback_checks_present",
-                    "status": "pass" if not rollback_check.missing_checks else "fail",
-                    "threshold": len(rollback_check.required_checks),
-                    "value": len(rollback_check.missing_checks),
-                }
-            ],
-            "observations": {
-                "required_checks": list(rollback_check.required_checks),
-                "missing_checks": list(rollback_check.missing_checks),
-            },
-            "reasons": list(rollback_check.reasons),
-            "artifact_refs": [str(output_dir / "gates" / "rollback-readiness.json")],
-        },
-        {
-            "name": "drift-gate",
-            "status": drift_gate_status,
-            "timestamp": phase_timestamp,
-            "slo_gates": [
-                {
-                    "id": "slo_drift_gate_allowed",
-                    "status": "pass" if drift_gate_check.get("allowed") else "fail",
-                    "threshold": True,
-                    "value": bool(drift_gate_check.get("allowed", False)),
-                }
-            ],
-            "observations": {
-                "eligible_for_live_promotion": bool(
-                    drift_gate_check.get("eligible_for_live_promotion", False)
-                ),
-                "drift_artifacts": drift_gate_artifacts,
-            },
-            "artifact_refs": drift_artifacts,
-            "reasons": list(drift_gate_check.get("reasons", [])),
-        },
-        {
-            "name": "paper-canary",
-            "status": canary_status,
-            "timestamp": phase_timestamp,
-            "slo_gates": [
-                {
-                    "id": "slo_paper_canary_patch_present",
-                    "status": canary_slo_status,
-                    "threshold": "patch exists for paper target",
-                    "value": str(patch_path) if patch_path else None,
-                }
-            ],
-            "observations": {
-                "target": requested_promotion_target,
-                "patch_path": str(patch_path) if patch_path else None,
-            },
-            "artifact_refs": ([str(patch_path)] if patch_path else []),
-        },
-        {
-            "name": "runtime-governance",
-            "status": runtime_gate_status,
-            "timestamp": phase_timestamp,
-            "slo_gates": [
-                {
-                    "id": "slo_runtime_rollback_not_triggered",
-                    "status": "pass" if not rollback_triggered else "fail",
-                    "threshold": False,
-                    "value": rollback_triggered,
-                }
-            ],
-            "observations": {
-                "requested_promotion_target": requested_promotion_target,
-                "drift_status": str(runtime_governance.get("drift_status", "unknown")),
-                "action_type": str(runtime_governance.get("action_type", "")) or None,
-                "action_triggered": bool(
-                    runtime_governance.get("action_triggered", False)
-                ),
-                "rollback_triggered": rollback_triggered,
-            },
-            "required": {"required_items": ["drift_status", "rollback_triggered"]},
-            "artifact_refs": runtime_artifact_refs,
-            "reasons": list(runtime_governance.get("reasons", [])),
-        },
-        {
-            "name": "rollback-proof",
-            "status": rollback_proof_status,
-            "timestamp": phase_timestamp,
-            "slo_gates": [
-                {
-                    "id": "slo_rollback_evidence_required_when_triggered",
-                    "status": (
-                        "pass"
-                        if rollback_proof_path and rollback_triggered
-                        else "pass"
-                        if not rollback_triggered
-                        else "fail"
-                    ),
-                    "threshold": True,
-                    "value": bool(rollback_proof_path),
-                }
-            ],
-            "observations": {
-                "rollback_triggered": rollback_triggered,
-                "rollback_incident_evidence_path": rollback_proof_path or "",
-                "rollback_incident_evidence": rollback_proof_path or "",
-            },
-            "artifact_refs": [rollback_proof_path] if rollback_proof_path else [],
-            "reasons": list(rollback_proof.get("reasons", [])),
-        },
-    ]
-
-    overall_pass = all(
-        coerce_phase_status(phase.get("status"), default="fail")
-        in {"pass", "skipped", "skip"}
-        for phase in phase_summaries
-    )
-
-    phase_artifact_refs = sorted(
-        {
-            str(item)
-            for phase in phase_summaries
-            for item in cast(list[Any], phase.get("artifact_refs", []))
-            if str(item).strip()
-        }
-    )
-    evidence_paths = sorted(
-        {
-            str(output_dir / "gates" / "profitability-benchmark-v4.json"),
-            str(output_dir / "gates" / "profitability-evidence-v4.json"),
-            str(output_dir / "gates" / "profitability-evidence-validation.json"),
-            str(output_dir / "gates" / "benchmark-parity-report-v1.json"),
-            str(output_dir / "gates" / "janus-event-car-v1.json"),
-            str(output_dir / "gates" / "janus-hgrm-reward-v1.json"),
-            str(output_dir / "gates" / "promotion-evidence-gate.json"),
-            str(output_dir / "gates" / "promotion-prerequisites.json"),
-            str(output_dir / "gates" / "rollback-readiness.json"),
-            str(output_dir / "research" / "candidate-spec.json"),
-            str(output_dir / "backtest" / "evaluation-report.json"),
-            str(output_dir / "backtest" / "walkforward-results.json"),
-            str(output_dir / "rollout" / "phase-manifest.json"),
-        }
-    )
-    phase_transitions = normalize_phase_transitions(
-        phase_summaries,
-    )
-
-    canonical_phase_payloads = {
-        str(phase.get("name", "")).strip(): dict(phase)
-        for phase in phase_summaries
-        if str(phase.get("name", "")).strip() in _AUTONOMY_PHASE_ORDER
-    }
-    ordered_phase_summaries: list[dict[str, Any]] = []
-    for expected_name in _AUTONOMY_PHASE_ORDER:
-        phase_payload = canonical_phase_payloads.get(expected_name)
-        if phase_payload is None:
-            ordered_phase_summaries.append(
-                {
-                    "name": expected_name,
-                    "status": "skip",
-                    "timestamp": phase_timestamp,
-                    "observations": {"note": "stage not evaluated"},
-                    "slo_gates": [],
-                    "artifact_refs": [],
-                }
-            )
-            continue
-
-        normalized_phase = dict(phase_payload)
-        normalized_phase["name"] = expected_name
-        normalized_phase["status"] = coerce_phase_status(
-            normalized_phase.get("status"), default="skip"
-        )
-        normalized_phase["artifact_refs"] = _coerce_path_strings(
-            normalized_phase.get("artifact_refs", [])
-        )
-        normalized_phase.setdefault("slo_gates", [])
-        ordered_phase_summaries.append(normalized_phase)
-
-    phase_summaries = ordered_phase_summaries
-    phase_transitions = normalize_phase_transitions(phase_summaries)
-    normalized_runtime_governance = dict(runtime_governance)
-    normalized_runtime_governance["rollback_triggered"] = rollback_triggered
-    normalized_runtime_governance["rollback_incident_evidence_path"] = (
-        rollback_proof_path
-    )
-    normalized_runtime_governance["rollback_incident_evidence"] = rollback_proof_path
-    normalized_rollback_proof = dict(rollback_proof)
-    normalized_rollback_proof["rollback_triggered"] = rollback_triggered
-    normalized_rollback_proof["rollback_incident_evidence_path"] = rollback_proof_path
-    normalized_rollback_proof["rollback_incident_evidence"] = rollback_proof_path
-
-    return {
-        "schema_version": "autonomy-phase-manifest-v1",
-        "run_id": run_id,
-        "candidate_id": candidate_id,
-        "execution_context": {
-            "repository": execution_context.get("repository", "unknown"),
-            "base": execution_context.get("base", "unknown"),
-            "head": execution_context.get("head", "unknown"),
-            "artifactPath": artifact_path,
-            "priorityId": execution_context.get("priorityId", ""),
-        },
-        "requested_promotion_target": requested_promotion_target,
-        "created_at": phase_timestamp,
-        "updated_at": phase_timestamp,
-        "phase_count": len(phase_summaries),
-        "phase_transitions": phase_transitions,
-        "status": "pass" if overall_pass else "fail",
-        "observation_summary": {
-            "signal_count": len(signals),
-            "has_signals": bool(signals),
-            "paper_canary_targeted": requested_promotion_target == "paper",
-            "live_targeted": requested_promotion_target == "live",
-        },
-        "phases": phase_summaries,
-        "runtime_governance": normalized_runtime_governance,
-        "rollback_proof": normalized_rollback_proof,
-        "artifact_refs": sorted(
-            {
-                artifact_ref
-                for artifact_ref in [*phase_artifact_refs, *evidence_paths]
-                if str(artifact_ref).strip()
-            }
-        ),
-        "slo_contract_version": "governance-slo-v1",
-    }
-
-
-def _coerce_str(raw: Any, default: str = "") -> str:
-    if not isinstance(raw, str):
-        return default
-    return raw.strip() or default
-
-
-def _coerce_int(raw: Any, default: int = 0) -> int:
-    if isinstance(raw, bool):
-        return int(raw)
-    if isinstance(raw, int):
-        return raw
-    if isinstance(raw, float):
-        return int(raw)
-    if isinstance(raw, str):
-        try:
-            return int(raw)
-        except ValueError:
-            return default
-    return default
-
-
-def _coerce_path_strings(values: Any) -> list[str]:
-    if not isinstance(values, (list, tuple, set)):
-        return []
-    value_items = cast(Iterable[object], values)
-    return sorted({_coerce_str(value) for value in value_items if _coerce_str(value)})
-
-
-def _coerce_gate_phase_gates(raw_gates: Any) -> list[dict[str, Any]]:
-    gates: list[dict[str, Any]] = []
-    if not isinstance(raw_gates, list):
-        return gates
-    for item in cast(list[object], raw_gates):
-        if not isinstance(item, dict):
-            continue
-        payload = cast(dict[str, Any], item)
-        gate_id = str(payload.get("gate_id", "")).strip()
-        if not gate_id:
-            continue
-        status = str(payload.get("status", "fail")).strip()
-        gates.append(
-            {
-                "id": gate_id,
-                "status": status if status else "fail",
-                "reasons": payload.get("reasons", []),
-                "value": payload.get("value"),
-                "threshold": payload.get("threshold"),
-            }
-        )
-    return gates
-
-
-def _build_actuation_intent_payload(
-    *,
-    run_id: str,
-    candidate_id: str,
-    generated_at: datetime,
-    recommendation_trace_id: str,
-    gate_report_trace_id: str,
-    promotion_target: str,
-    recommended_mode: str,
-    actuation_allowed: bool,
-    promotion_check: dict[str, Any],
-    rollback_check: dict[str, object],
-    candidate_state_payload: dict[str, Any],
-    gate_report_path: Path,
-    rollback_check_path: Path,
-    candidate_spec_path: Path,
-    candidate_generation_manifest_path: Path,
-    evaluation_manifest_path: Path,
-    recommendation_manifest_path: Path,
-    profitability_manifest_path: Path,
-    promotion_recommendation_path: Path,
-    evaluation_report_path: Path,
-    walk_results_path: Path,
-    paper_patch_path: Path | None,
-    patch_required: bool,
-    profitability_benchmark_path: Path,
-    profitability_evidence_path: Path,
-    profitability_validation_path: Path,
-    simulation_calibration_report_path: Path,
-    shadow_live_deviation_report_path: Path,
-    benchmark_parity_path: Path,
-    foundation_router_parity_path: Path,
-    deeplob_bdlob_report_path: Path,
-    advisor_fallback_slo_report_path: Path,
-    janus_event_car_path: Path,
-    janus_hgrm_reward_path: Path,
-    recalibration_report_path: Path,
-    promotion_gate_path: Path,
-    promotion_recommendation: PromotionRecommendation,
-    recommendations: list[str],
-    governance_repository: str,
-    governance_base: str,
-    governance_head: str,
-    governance_artifact_path: str,
-    priority_id: str | None,
-    governance_change: str,
-    governance_reason: str,
-    stage_lineage_payload: dict[str, Any],
-    replay_artifact_hashes: dict[str, str],
-    candidate_hash: str,
-) -> dict[str, Any]:
-    raw_missing_checks = cast(object, rollback_check.get("missing_checks"))
-    rollback_missing_checks = (
-        list(cast(list[object], raw_missing_checks))
-        if isinstance(raw_missing_checks, list)
-        else []
-    )
-    rollback_evidence_links: list[str] = []
-    rollback_evidence_links.extend(
-        [
-            str(candidate_spec_path),
-            str(gate_report_path),
-            str(evaluation_report_path),
-            str(walk_results_path),
-            str(promotion_gate_path),
-            str(candidate_generation_manifest_path),
-            str(evaluation_manifest_path),
-            str(recommendation_manifest_path),
-            str(profitability_manifest_path),
-            str(promotion_recommendation_path),
-            str(profitability_benchmark_path),
-            str(benchmark_parity_path),
-            str(foundation_router_parity_path),
-            str(deeplob_bdlob_report_path),
-            str(advisor_fallback_slo_report_path),
-            str(profitability_evidence_path),
-            str(profitability_validation_path),
-            str(simulation_calibration_report_path),
-            str(shadow_live_deviation_report_path),
-            str(janus_event_car_path),
-            str(janus_hgrm_reward_path),
-            str(recalibration_report_path),
-            str(rollback_check_path),
-        ]
-    )
-    if paper_patch_path is not None:
-        rollback_evidence_links.append(str(paper_patch_path))
-    rollback_evidence_links.extend(
-        [str(item) for item in promotion_check.get("artifact_refs", [])]
-    )
-    candidate_state_readiness = _candidate_state_readiness_payload(
-        candidate_state_payload
-    )
-    return {
-        "schema_version": _ACTUATION_INTENT_SCHEMA_VERSION,
-        "run_id": run_id,
-        "candidate_id": candidate_id,
-        "generated_at": generated_at.isoformat(),
-        "generated_for": "autonomous-lane",
-        "promotion_target": promotion_target,
-        "recommended_mode": recommended_mode,
-        "governance": {
-            "repository": governance_repository,
-            "base": governance_base,
-            "head": governance_head,
-            "artifact_path": governance_artifact_path,
-            "change": governance_change,
-            "reason": governance_reason,
-            "priority_id": priority_id,
-        },
-        "actuation_allowed": actuation_allowed,
-        "confirmation_phrase_required": promotion_target == "live",
-        "confirmation_phrase": (
-            _ACTUATION_CONFIRMATION_PHRASE if promotion_target == "live" else None
-        ),
-        "gates": {
-            "recommendation_trace_id": recommendation_trace_id,
-            "gate_report_trace_id": gate_report_trace_id,
-            "promotion_allowed": promotion_recommendation.eligible,
-            "promotion_action": promotion_recommendation.action,
-            "recommendation_reasons": recommendations,
-        },
-        "candidate_hash": candidate_hash,
-        "artifact_refs": sorted(
-            {item for item in rollback_evidence_links if item.strip()}
-        ),
-        "audit": {
-            "candidate_state_payload": candidate_state_payload,
-            "promotion_check": promotion_check,
-            "rollback_check": rollback_check,
-            "stage_lineage": stage_lineage_payload,
-            "replay_artifact_hashes": replay_artifact_hashes,
-            "patch_required": patch_required,
-            "patch_available": paper_patch_path is not None,
-            "rollback_readiness_readout": {
-                "kill_switch_dry_run_passed": bool(
-                    candidate_state_readiness.get("killSwitchDryRunPassed")
-                ),
-                "gitops_revert_dry_run_passed": bool(
-                    candidate_state_readiness.get("gitopsRevertDryRunPassed")
-                ),
-                "strategy_disable_dry_run_passed": bool(
-                    candidate_state_readiness.get("strategyDisableDryRunPassed")
-                ),
-                "human_approved": bool(candidate_state_readiness.get("humanApproved")),
-                "rollback_target": str(
-                    candidate_state_readiness.get("rollbackTarget") or ""
-                ),
-                "dry_run_completed_at": str(
-                    candidate_state_readiness.get("dryRunCompletedAt") or ""
-                ),
-            },
-            "rollback_evidence_missing_checks": rollback_missing_checks,
-        },
-    }
-
-
-def _candidate_state_readiness_payload(
-    candidate_state_payload: dict[str, Any],
-) -> dict[str, Any]:
-    candidate_state_readiness_raw = candidate_state_payload.get("rollbackReadiness", {})
-    return (
-        cast(dict[str, Any], candidate_state_readiness_raw)
-        if isinstance(candidate_state_readiness_raw, dict)
-        else {}
-    )
 
 
 def _collect_walk_decisions_for_runtime(
@@ -7625,10 +5127,6 @@ def _baseline_runtime_strategies() -> list[StrategyRuntimeConfig]:
     ]
 
 
-def _sha256_path(path: Path) -> str:
-    return hashlib.sha256(path.read_bytes()).hexdigest()
-
-
 def _profitability_threshold_payload(
     policy_payload: dict[str, Any],
 ) -> dict[str, object]:
@@ -7670,25 +5168,6 @@ def _collect_confidence_values(decisions: list[WalkForwardDecision]) -> list[Dec
         except (ArithmeticError, TypeError, ValueError):
             continue
     return values
-
-
-def _default_strategy_configmap_path() -> Path:
-    for parent in Path(__file__).resolve().parents:
-        if (parent / "argocd").is_dir() and (parent / "services" / "torghut").is_dir():
-            return (
-                parent
-                / "argocd"
-                / "applications"
-                / "torghut"
-                / "strategy-configmap.yaml"
-            )
-    return (
-        Path(__file__).resolve().parents[6]
-        / "argocd"
-        / "applications"
-        / "torghut"
-        / "strategy-configmap.yaml"
-    )
 
 
 def _to_orm_strategies(
