+++ services/torghut/tests/test_explicit_module_facade_exports.py
@@ -19,6 +19,13 @@ EXPLICIT_IMPORT_SURFACES = (
     "app.models.entities",
     "app.trading.alpha.lane",
     "app.trading.autonomy.gates",
+    "app.trading.autonomy.lane_candidate_payloads",
+    "app.trading.autonomy.lane_common",
+    "app.trading.autonomy.lane_phase_payloads",
+    "app.trading.autonomy.lane_profitability_manifest",
+    "app.trading.autonomy.lane_regime_artifacts",
+    "app.trading.autonomy.lane_stage_artifacts",
+    "app.trading.autonomy.lane_strategy_factory",
     "app.trading.autonomy.policy_check",
     "app.trading.completion",
     "app.trading.decisions",
